% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{acc}
\alias{acc}
\title{Calculates accuracy from a measures dataframe}
\usage{
acc(df)
}
\arguments{
\item{df}{dataframe with measures}
}
\value{
calculated metric
}
\description{
Calculates accuracy from a measures dataframe
}
\examples{
df <- data.frame(TP = 4, FP = 3, TN = 2, FN = 1)
acc(df)
}
\keyword{metrics}
