% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{add_exp_design}
\alias{add_exp_design}
\title{add_exp_design
Add experimental condition to Seurat metadata}
\usage{
add_exp_design(seu, name, exp_design)
}
\arguments{
\item{seu}{Seurat object}

\item{name}{Sample name}

\item{exp_design}{Experiment design table in CSV format}
}
\value{
Seurat object with the experimental conditions added as metadata
}
\description{
add_exp_design
Add experimental condition to Seurat metadata
}
\keyword{integration}
\keyword{preprocessing,}
\keyword{subsetting,}
