% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{AE_Gene_Overview}
\alias{AE_Gene_Overview}
\title{Simple wrapper for OUTRIDER::plotExpressionRank and
plotExpectedVsObservedCounts functions. Calls them for specified
genes in input Outrider DataSet (ods)}
\usage{
AE_Gene_Overview(gene)
}
\arguments{
\item{gene}{Gene to plot.}
}
\value{
Volcano plots of provided gene expression rank and expected vs
observed counts across all samples.
}
\description{
\code{AE_Gene_Overview} takes a gene and an Outrider DataSet (ODS), and
prints two plots. The first one is the Expression Rank, which represents
normalized counts of specified gene by sample. Samples are sorted in
ascending order by normalized counts (i.e expression rank). Outliers are
coloured red. ExpectedVsObserved plot creates a graph which represents
raw (observed) versus expected counts. Expected counts are an estimation
based on OUTRIDER normalization and total expression levels in sample.
}
