% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_expression_packages.R
\name{analysis_DESeq2}
\alias{analysis_DESeq2}
\title{analysis_DESeq2}
\usage{
analysis_DESeq2(
  data,
  p_val_cutoff,
  target,
  model_formula_text,
  multifactorial,
  var_data = NULL
)
}
\arguments{
\item{data}{Data to compute}

\item{p_val_cutoff}{P-value threshold to mark a DEG as significant}

\item{target}{Data frame describing sample grouping for DE analysis}

\item{model_formula_text}{design formula used to specify the statistical
model used in the DE analysis step. Generated automatically in our workflow}

\item{multifactorial}{String describing multifactorial analysis to perform}

\item{var_data}{Pre-computed DESeq2 data run with default parameters. If
DESeq2 is called with default parameters, this dds object will be used
instead of calculating it again.}
}
\value{
A list of two objects: de_deseq2, which contains the DESeq2 DE
analysis, and DESeq2_dataset, which contains only the DESeq dataset.
}
\description{
analysis_DESeq2
}
