% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{analyze_query}
\alias{analyze_query}
\title{analyze_query
\code{analyze_query} is a wrapper for the three query analysis steps:
\code{get_query_distribution}, \code{get_query_pct} by samples and \code{get_query_pct} by
samples and cell types.}
\usage{
analyze_query(seu, query, sigfig)
}
\arguments{
\item{seu}{Seurat object}

\item{query}{Vector of query genes whose expression to analyse.}

\item{sigfig}{Significant figure cutoff}
}
\value{
A list with the three query analysis objects.
}
\description{
analyze_query
\code{analyze_query} is a wrapper for the three query analysis steps:
\code{get_query_distribution}, \code{get_query_pct} by samples and \code{get_query_pct} by
samples and cell types.
}
