% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{annotate_SingleR}
\alias{annotate_SingleR}
\title{annotate_SingleR}
\usage{
annotate_SingleR(
  seu,
  SingleR_ref = NULL,
  ref_n = 25,
  integrate = "TRUE",
  ref_de_method = "wilcox",
  BPPARAM = BiocParallel::SerialParam(),
  ref_label = ref_label,
  aggr.ref = FALSE,
  fine.tune = TRUE,
  assay = "RNA",
  verbose = FALSE,
  subset_by = NULL
)
}
\arguments{
\item{seu}{A seurat object.}

\item{SingleR_ref}{SummarizedExperiment object to use as reference
for SingleR cell type annotation. If NULL (default value),
SingleR will not be used.}

\item{ref_n}{Top N reference markers to consider in annotation. Higher values
provide a more accurate annotation, but increase noise and computational
time. Will not be used if ref_de_method is empty.}

\item{integrate}{A boolean.
\itemize{
\item \code{TRUE}: Integrate seurat object using the \code{harmony} package.
\item \code{FALSE} (the default): Do not perform integration.
}}

\item{ref_de_method}{Method to use for marker calculation in single-cell
reference.}

\item{BPPARAM}{Parameters to pass to BiocParallel framework.}

\item{ref_label}{Reference column to use in annotation.}

\item{aggr.ref}{Logical scalar indicating whether references should be aggregated to pseudo-bulk samples for speed, see \code{\link[SingleR]{aggregateReference}}.}

\item{fine.tune}{A boolean.
\itemize{
\item \code{TRUE} (the default): Fine-tune statistical model built for transfer.
\item \code{FALSE}: Do not perform fine-tuning..
}}

\item{assay}{Seurat object assay to use for annotation. Default "RNA".}

\item{verbose}{A boolean.
\itemize{
\item \code{TRUE}: Prints progress bars and messages to closely monitor progress.
\item \code{FALSE} (the default): Print minimal progress messages.
}}

\item{subset_by}{A string vector. Categories to consider in integrative
analysis.}
}
\value{
A seurat object with a new sketched assay.
}
\description{
\code{annotate_SingleR} is a wrapper for SingleR annotation strategy,
simplifies function call.
}
\examples{
 \dontrun{
   annotate_SingleR(seu = seu, SingleR_ref = SingleR_ref,
   BPPARAM = BPPARAM, ref_n = ref_n, ref_label = ref_label,
   verbose = verbose, ref_de_method = ref_de_method, aggr.ref = aggr.ref,
   fine.tune = new_opt$fine.tune)
 }
}
