% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{catched_pairwise_termsim}
\alias{catched_pairwise_termsim}
\title{Catched errors fo pairwise_termsim for special cases}
\usage{
catched_pairwise_termsim(enr, num_cats = 200)
}
\arguments{
\item{enr}{enrichment object to be studied}

\item{num_cats}{number of categories to be shown}
}
\value{
enrichment object after add termsim info
}
\description{
Catched errors fo pairwise_termsim for special cases
}
