% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{check_results}
\alias{check_results}
\title{Check if a given object is a correct enrichment result object}
\usage{
check_results(df)
}
\arguments{
\item{df}{variable to be checked}
}
\value{
true if it is an allowed object or false in other cases
}
\description{
Check if a given object is a correct enrichment result object
}
\keyword{check}
