% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{check_results_list}
\alias{check_results_list}
\title{Check if a given list of objects is a correct enrichment result objects list}
\usage{
check_results_list(ls)
}
\arguments{
\item{ls}{variable to be checked}
}
\value{
true if it is an allowed object or false in other cases
}
\description{
Check if a given list of objects is a correct enrichment result objects list
}
\keyword{check}
