% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{collapse_markers}
\alias{collapse_markers}
\title{collapse_markers
\code{collapse_markers} takes list of marker gene data frames and collapses it
into a cluster-markers data frame.}
\usage{
collapse_markers(markers_list)
}
\arguments{
\item{markers_list}{A list containing marker gene data frames.}
}
\value{
A data frame. Column \code{cluster} contains element names of original
list (seurat clusters) and column \code{genes} contains, for each row, the top
markers of that element from the original list separated by commas.
}
\description{
collapse_markers
\code{collapse_markers} takes list of marker gene data frames and collapses it
into a cluster-markers data frame.
}
\examples{
 \dontrun{
   collapse_markers(markers_list = list_of_marker_dfs)
 }
}
