% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_packages.R
\name{corM2igraph}
\alias{corM2igraph}
\title{corM2igraph}
\usage{
corM2igraph(corM, cor_abs_thr = 0.75)
}
\arguments{
\item{corM}{correlation matrix.}

\item{cor_abs_thr}{Correlation threshold (absolute value)}
}
\value{
An Igraph.'
}
\description{
\code{corM2igraph} Creates an igraph out of a correlation matrix.
}
\examples{
 \dontrun{
     corM2igraph(matrix, 0.75)
 }
}
