% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cormit.R
\name{coRmiT}
\alias{coRmiT}
\title{Main coRmiT Function}
\usage{
coRmiT(
  RNAseq_folder,
  miRNAseq_folder,
  output_files,
  strat_names,
  sample_proportion,
  organism,
  multimir_db,
  p_val_cutoff,
  corr_cutoffs,
  permutations,
  MM_cutoff,
  report,
  translation_file,
  databases,
  translate_ensembl = FALSE,
  filter_db_theshold = 0,
  database_to_filter = NULL,
  mc_cores = 1,
  tag_filter,
  corr_type,
  corr_coef,
  f_p_val,
  selected_targets_file,
  eval_method = "aggregated",
  template_folder = file.path(find.package("ExpHunterSuite"), "templates"),
  organism_table_path = file.path(find.package("ExpHunterSuite"), "inst",
    "external_data", "organism_table.txt"),
  compare_pred_scores = FALSE,
  add_databases_files = NA
)
}
\arguments{
\item{RNAseq_folder}{Output expression hunter folder for mRNAs}

\item{miRNAseq_folder}{Output expression hunter folder for miRNAs}

\item{output_files}{Output folder}

\item{strat_names}{Strategies to employ to find miRNA-target pairs}

\item{sample_proportion}{Score distribution sample proportion}

\item{organism}{Reference organism to use. hsa for human, mmu mouse}

\item{multimir_db}{Indicate .RData file with parsed multiMiR data}

\item{p_val_cutoff}{Correlation P value threshold}

\item{corr_cutoffs}{Correlation thresholds to test for}

\item{permutations}{Permutations of random tests}

\item{MM_cutoff}{Genes with lower module membership than this remove}

\item{report}{Name of the html file}

\item{translation_file}{Two columns file with miRNA translations
Same IDs as input in first column, miRBase ID in second e.g. MIMAT000000}

\item{databases}{prediction databases included from multiMiR for GS}

\item{translate_ensembl}{Translate mRNA Ensembl ID to ENTREZ & GENESYMBOL}

\item{filter_db_theshold}{Minimun prediction databases to support a pair}

\item{database_to_filter}{Predicted databases that must support a pair on}

\item{mc_cores}{cores to use for parallelization}

\item{tag_filter}{filter type for RNAseq and miRNAseq}

\item{corr_type}{Find pairs with correlation lower or higher than corr cutoff}

\item{corr_coef}{Correlation method: pearson, spearman, kendall, bicor}

\item{f_p_val}{Fisher P value threshold for overlap with multiMiR}

\item{selected_targets_file}{One column file indicating external selected targets}

\item{eval_method}{Apply odds ratio etc. to all pairs (aggregated)
or miRNA by miRNA (specific)}

\item{template_folder}{where to find templates for the report}

\item{organism_table_path}{table with data for organism used}

\item{compare_pred_scores}{Compute predition scores comparison}

\item{add_databases_files}{Use additional target DBs (gene \if{html}{\out{<tab>}} miRNA)}
}
\value{
folder with miRNA-target pairs, report and other details.
}
\description{
This function allows you to detect mirna-target pairs
}
\examples{
\dontrun{
 coRmiT(RNAseq_folder = "mirna_path", miRNAseq_folder="mirna_path",
   strat_names = c("EE","Eh","Ed"), --organism = "hsa",
   corr_cutoffs = "-0.95,-0.9")
}
}
