% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{create_DEG_lists}
\alias{create_DEG_lists}
\title{Randomly select which genes will be differentially expressed
\code{create_DEG_lists} takes a list of genes and chooses which among them will
be differentially expressed.}
\usage{
create_DEG_lists(universe, nDEGs, overlap_size)
}
\arguments{
\item{universe}{List of genes from which DEGs will be chosen}

\item{nDEGs}{Vector of number of DEGs to select. One number for every list
of DEGs to generate}

\item{overlap_size}{Proportion of DEGs that multiple experimental conditions
will share}
}
\value{
A list containing a vector of DEGs for every experimental condition
}
\description{
Randomly select which genes will be differentially expressed
\code{create_DEG_lists} takes a list of genes and chooses which among them will
be differentially expressed.
}
