% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{custom_synth}
\alias{custom_synth}
\title{custom_synth}
\usage{
custom_synth(
  counts_table,
  exp_design,
  nDEGs,
  columns,
  fixed_DEG_lists,
  fixed_upregulated_DEGs,
  effect_sizes,
  fraction_upregulated,
  overlap_size,
  deg_method,
  samples_per_cond
)
}
\arguments{
\item{counts_table}{A data frame. Counts table to manipulate for DEG
synthesis}

\item{exp_design}{Experimental design}

\item{nDEGs}{Vector of number of DEGs to select. One number for every list
of DEGs to generate}

\item{columns}{An integer vector. Columns to use as factors for DEG synthesis}

\item{fixed_DEG_lists}{A string. Comma-separated paths to DEG lists to use.
Will disable random gene selection}

\item{fixed_upregulated_DEGs}{A string. Comma-separated paths to upregulated
DEGs list. will disable random upregulated DEG selection}

\item{effect_sizes}{A string. Comma-separated effect sizes to use for each
factor}

\item{fraction_upregulated}{A numeric. Fraction of DEGs to upregulate}

\item{overlap_size}{Proportion of DEGs that multiple experimental conditions
will share}

\item{deg_method}{Method to use. Values: "norm" (\code{synth_normal_DEGs}), det
(\code{synth_deterministic_DEGs}) and "exp" (\code{synth_exponential_DEGs})}

\item{samples_per_cond}{An integer. Number of samples that each group
will have}
}
\value{
A list. Element \code{all} contains the entire results object. Element
\code{counts_matrix} contains just the counts matrix. Element \code{DEGs} contains
a data frame detailing which genes are differentially expressed in dataset
}
\description{
\code{custom_synth} Is a wrapper for the \code{generateSyntheticData} function
from package \code{compcodeR}.
}
\examples{
B_625_625 <- custom_synth(dataset = "B_625_625", n.vars = 12500, repl.id = 1,
                 samples.per.cond = 5, n.diffexp = 1250, seqdepth = 1e7, 
                 fraction.upregulated = 0.5, between.group.diffdisp = FALSE,
                 filter.threshold.total = 1, filter.threshold.mediancpm = 0,
                 fraction.non.overdispersed = 0)
B_625_625
}
