% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{CV}
\alias{CV}
\title{Calculate the coefficient of variation in a vector
\code{CV} performs a simple calculation to retrieve the coefficient of variation
of a vector}
\usage{
CV(vector)
}
\arguments{
\item{vector}{Vector whose CV to calculate}
}
\value{
The coefficient of variation of the vector
}
\description{
Calculate the coefficient of variation in a vector
\code{CV} performs a simple calculation to retrieve the coefficient of variation
of a vector
}
