% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{define_conditions}
\alias{define_conditions}
\title{Extract samples by their value of requested experimental condition.
\code{define_conditions} takes an experimental design and the index of a factor
column and creates two vectors, one for each value of the supplied condition}
\usage{
define_conditions(exp_design, factor_column)
}
\arguments{
\item{exp_design}{Experimental design}

\item{factor_column}{Index of column whose factor to consider}
}
\value{
Two vectors of samples, one for each factor.
}
\description{
Extract samples by their value of requested experimental condition.
\code{define_conditions} takes an experimental design and the index of a factor
column and creates two vectors, one for each value of the supplied condition
}
