% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degh_output}
\alias{degh_output}
\title{Output data from the main_degenes_Hunter function}
\format{
An output object from the main_degenes_Hunter function.
}
\usage{
data(degh_output)
}
\description{
Output data from the main_degenes_Hunter function
}
\examples{
data(degh_output)
}
\keyword{datasets}
