% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{DESeq4MAE}
\alias{DESeq4MAE}
\title{Run deseq test for MAE}
\usage{
DESeq4MAE(data, minCoverage = 10, disp = 0.05, independentFiltering = FALSE)
}
\arguments{
\item{data}{A data.frame containing allelic counts.}

\item{minCoverage}{minimum total allelic count. Default is 10.}

\item{disp}{Gene dispersion for the NB test. Default is 0.05.}

\item{independentFiltering}{Parameter that affects the multiple testing. Default is FALSE.}
}
\value{
Mono-allelic results table containing original counts plus p-value, p-adjusted and freqALT columns.
}
\description{
Uses a negative binomial test to determine if a variant is mono-allelically expressed.
}
\examples{
file <- system.file("extdata", "allelic_counts_HG00187.csv", package = "tMAE", mustWork = TRUE)
maeCounts <- fread(file)
DESeq4MAE(maeCounts)
}
\author{
Vicente Yepez, Christian Mertes
}
