% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{.add_base_IDs}
\alias{.add_base_IDs}
\title{Add base ID column to sample annotation data frame}
\usage{
.add_base_IDs(col_data)
}
\arguments{
\item{col_data}{Data frame containing RNA_ID column.}
}
\value{
The same dataframe, with a new BASE_ID column containing sample ID.
If RNA_ID column only contains sample ID, new column will be an exact copy.
}
\description{
\code{.add_base_IDs} takes a sample annotation data frame and calls .get_base_ID
on it, then assigns the results to new BASE_ID column.
}
