% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{.add_HPO_cols}
\alias{.add_HPO_cols}
\title{Add HPO terms to table}
\usage{
.add_HPO_cols(
  RES,
  sample_id_col = "sampleID",
  gene_name_col = "hgncSymbol",
  hpo_file = NULL
)
}
\arguments{
\item{RES}{Data table to process.}

\item{sample_id_col}{Name of sample id column.}

\item{gene_name_col}{Name of column where hgnc symbols will be included.}

\item{hpo_file}{File containing HPO terms associated to genes.}
}
\value{
Data frame with new hgnc symbol column and (optionally) HPO terms
associated to symbols.
}
\description{
\code{.add_HPO_cols} takes a data table and adds a hgncSymbol column.
}
