% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{.fix_lapply_names}
\alias{.fix_lapply_names}
\title{Function to fix an annoying characteristic of using lapply, where it
appends the original object name to every name of the result list.
\code{.fix_lapply_names} Fixes the output of a lapply call, restoring the original
names of the object it was called upon.}
\usage{
.fix_lapply_names(table_list)
}
\arguments{
\item{table_list}{A list of tables originating from a lapply call.}
}
\value{
That same list with the expected names, without the appends made
by the lapply call.
}
\description{
Function to fix an annoying characteristic of using lapply, where it
appends the original object name to every name of the result list.
\code{.fix_lapply_names} Fixes the output of a lapply call, restoring the original
names of the object it was called upon.
}
