% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{.get_clusters_df}
\alias{.get_clusters_df}
\title{Clusterize correlation data.
\code{.get_clusters_df} Clusterizes a count correlation matrix, adding metadata
specified in the ods object from which it was calculated.}
\usage{
.get_clusters_df(
  ods,
  matrix,
  groups = character(),
  nClust = 4,
  type = "sample"
)
}
\arguments{
\item{ods}{Outrider DataSet object.}

\item{matrix}{log2 counts correlation matrix.}

\item{groups}{A string. Column of ods colData to use as secondary grouping
atribute. Default: "EXTERNAL".}

\item{nClust}{Number of clusters in which data should be divided.}

\item{type}{A string. "sample" to clusterize sample correlation data, "gene"
for genes.}
}
\value{
A data frame. Rows are named after samples. Columns are nClust
(cluster where sample belongs) and specified groups column.
}
\description{
Clusterize correlation data.
\code{.get_clusters_df} Clusterizes a count correlation matrix, adding metadata
specified in the ods object from which it was calculated.
}
