% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{.get_union_FCs}
\alias{.get_union_FCs}
\title{.get_union_FCs}
\usage{
.get_union_FCs(DEG_list, top = NA, min_log2FC = 0.5, p_val_cutoff = 0.05)
}
\arguments{
\item{DEG_list}{A list containing DEG results.}

\item{top}{An integer. Top DEGs to extract from each cell type, by logFC
value. Takes precedence over min_log2FC value. Default NULL.}

\item{min_log2FC}{A numeric. A log2FC threshold to set for top DEG selection.
default 2.}

\item{p_val_cutoff}{Maximum p-value to consider a genes as differentially
expressed.}
}
\value{
A heatmap containing the log2FC value for the union of the top
DEGs of every cell type in the experiment, for every cell type in the
experiment.
}
\description{
\code{get_union_FCs} extracts the union of all differentially expressed genes
across every cell type in a single-cell experiment, then returns a heatmap
that shows its FC value in every cell type. It is a specific cell type
plot, so global DEGs are excluded from the calculation.
}
