% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{.map_genes}
\alias{.map_genes}
\title{Create gene name mapping file}
\usage{
.map_genes(gtf)
}
\arguments{
\item{gtf}{GTF file to process.}
}
\value{
A data frame containing GTF file gene annotation.
}
\description{
\code{map_genes} takes a GTF file and builds a gene-name mapping data frame out
of it.
}
