% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{.split_string_by_char}
\alias{.split_string_by_char}
\title{Function to split a string by a certain character and return only the desired
element.
\code{.split_string_by_char} Splits a string by the provided character, and
keeps the result specified by the provided index.}
\usage{
.split_string_by_char(string, char, index)
}
\arguments{
\item{string}{String to split.}

\item{char}{Char by which the string will be split.}

\item{index}{Index of the element after splitting that will be kept}
}
\value{
The element of the split string specified by provided index.
}
\description{
Function to split a string by a certain character and return only the desired
element.
\code{.split_string_by_char} Splits a string by the provided character, and
keeps the result specified by the provided index.
}
