% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{download_latest_kegg_db}
\alias{download_latest_kegg_db}
\title{Download KEGG db for a given organism}
\usage{
download_latest_kegg_db(current_organism_info, file)
}
\arguments{
\item{current_organism_info}{organism info for which to download the file}

\item{file}{where to save the file}
}
\value{
The latest KEGG db as a GSON object
}
\description{
Download KEGG db for a given organism
}
