% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{downsample_seurat}
\alias{downsample_seurat}
\title{Downsample a seurat object}
\usage{
downsample_seurat(
  seu,
  cells = 500,
  features = 5000,
  assay = "RNA",
  layer = "counts",
  keep = NULL
)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{cells, features}{Lists or integers. An integer will trigger random
downsampling by its respective variable.}

\item{assay}{Seurat object assay to break down.}

\item{layer}{Layer of seurat object assay to break down.}

\item{keep}{A string vector. Features to keep when downsampling.}
}
\value{
A downsampled seurat object.
}
\description{
\code{downsample_seurat} takes a seurat object as input, and downsamples it to
specified number of cells and features. You can also input specific lists
if you know which cells and/or features you want to retrieve.
}
\examples{
data(pbmc_tiny)
print(pbmc_tiny)
print(downsample_seurat(seu = pbmc_tiny, cells = 2, features = 5))
}
