% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{enrich_all_customs}
\alias{enrich_all_customs}
\title{Performs several CUSTOM enrichments using ORA method. You can give universes or load GMT files}
\usage{
enrich_all_customs(
  custom_sets = NULL,
  custom_files = NULL,
  p_val_threshold = 0.05,
  genes,
  write_path = NULL
)
}
\arguments{
\item{custom_sets}{custom universes to be used.}

\item{custom_files}{custom files to load universes to be used}

\item{p_val_threshold}{p-value threshold}

\item{genes}{significant genes to be used}

\item{write_path}{optional output path where enrichment tables will be stored}
}
\value{
list with enrichment tables obtained
}
\description{
Performs several CUSTOM enrichments using ORA method. You can give universes or load GMT files
}
\examples{
enrich_all_customs()
}
\keyword{enrich}
