% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{enrich_clusters_with_gmt}
\alias{enrich_clusters_with_gmt}
\title{Performs custom enrichments over cluster sets}
\usage{
enrich_clusters_with_gmt(
  custom_set,
  genes_in_modules = NULL,
  p_val_threshold,
  cores = 1,
  task_size = 1
)
}
\arguments{
\item{custom_set}{custom unvierses}

\item{genes_in_modules}{list of genes (clusters)}

\item{p_val_threshold}{p-value threshold}

\item{cores}{optional parallel cores to be used. Default: 1}

\item{task_size}{number of elements per packages used}
}
\value{
enrichment tables obtained
}
\description{
Performs custom enrichments over cluster sets
}
\examples{
# Will return NULL
enrich_clusters_with_gmt()
}
\keyword{enrich}
