% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{enrich_cnet}
\alias{enrich_cnet}
\alias{enrich_emap}
\alias{enrich_dotplot}
\title{Create ClusterProfiler plots}
\usage{
enrich_cnet(
  input_obj,
  n_category = 30,
  node_label = "all",
  layout = layout_nicely,
  size_category = 1,
  size_edge = 1,
  hilight = "none",
  hilight_alpha = 1,
  attr_vector = NULL,
  gene_attribute_name = NULL,
  advanced_opt = NULL,
  label_size = 2,
  legend_pos = "side"
)

enrich_emap(
  input_obj,
  n_category = 30,
  size_category = 1,
  layout = igraph::layout_nicely,
  label_format = 4,
  size_edge = 1,
  advanced_opt = NULL,
  label_size = 2.5,
  legend_pos = "side",
  node_label = "group"
)

enrich_dotplot(input_obj, n_category = 30)
}
\arguments{
\item{input_obj}{Enrichment results object}

\item{n_category}{An integer. Number of categories to pass as showCategory
value to several clusterProfiler plotting functions.}

\item{node_label}{How to label the nodes. Default "category". Other possible
values: "all", "none", "item", "exclusive", "share".}

\item{layout}{igraph layout to use}

\item{size_category}{Size of category labels. Smaller than 1 values will
reduce size, higher than 1 values will increase it. Default: 1.}

\item{size_edge}{Size of edges. Smaller than 1 values will reduce size,+
higher than 1 values will increase it. Default: 1.}

\item{hilight}{Name of category to highlight. Default "none".}

\item{hilight_alpha}{Transparency value that will be applied to categories
not highlighted. Must be a value between 0 (completely transparent) and 1
(completely opaque). Default: 0.3.}

\item{attr_vector}{Vector with extra attributes to include}

\item{gene_attribute_name}{Name of attribute vector}

\item{advanced_opt}{Named list of advanced options}

\item{label_size}{Font size for label}

\item{legend_pos}{A string. Where to put the legend. "bottom" to place it at
the bottom. Different value to place it at the side.}

\item{label_format}{An integer. Wrap length in characters}
}
\value{
Enriched plot
}
\description{
\code{enrich_cnet} takes a ClusterProfiler enriched object and creates a
concept network plot
}
