% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{enrich_density}
\alias{enrich_density}
\title{\code{enrich_density} draws a density plot for enrichment results.}
\usage{
enrich_density(enrich_result, attributes, showCategory = 30)
}
\arguments{
\item{enrich_result}{Enrichment result object.}

\item{attributes}{Enrichment attributes to retrieve from enrich_result.}

\item{showCategory}{showCategory. Default 30.}
}
\value{
Density plot of enrichment.
}
\description{
\code{enrich_density} draws a density plot for enrichment results.
}
\examples{
\dontrun{
   enrich_density(enrich_result, attrbutes = attribute_list)
}
}
