% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{estimateThetaWithoutAutoCorrect}
\alias{estimateThetaWithoutAutoCorrect}
\title{Estimate BCV (Biological Coefficient of Variation) from Outrider Dataset.}
\usage{
estimateThetaWithoutAutoCorrect(ods)
}
\arguments{
\item{ods}{Outrider DataSet object.}
}
\value{
An updated ODS containing estimated BCV.
}
\description{
\code{estimateThetaWithoutAutoCorrect} takes an ODS (Outrider DataSet) object
and calculates its biological coefficient of variation, adding it to the ODS.
}
