% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{exists_enrich_df}
\alias{exists_enrich_df}
\title{Function used to check if a specific result has been calculated}
\usage{
exists_enrich_df(var_name, level_depth = 3)
}
\arguments{
\item{var_name}{variable name to be checked}

\item{level_depth}{variable depth into a list of objects}
}
\value{
true if variable exists or false in other cases
}
\description{
Function used to check if a specific result has been calculated
}
\keyword{check}
