% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{exists_graph_file}
\alias{exists_graph_file}
\title{Checks  if a graph file exists}
\usage{
exists_graph_file(path, file_name, ext = "pdf")
}
\arguments{
\item{path}{to file}

\item{file_name}{file name}

\item{ext}{file extension}
}
\value{
true if file exists or false in other cases
}
\description{
Checks  if a graph file exists
}
\keyword{file}
