% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{extract_from_cm}
\alias{extract_from_cm}
\title{Extract value from confusion matrix}
\usage{
extract_from_cm(cm, real, pred, default = 0)
}
\arguments{
\item{cm}{confusion matrix}

\item{real}{slot name}

\item{pred}{slot name}

\item{default}{to return if value is not available}
}
\value{
value required or default if it is not available
}
\description{
Extract value from confusion matrix
}
