% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata from seurat object.}
\usage{
extract_metadata(seu)
}
\arguments{
\item{seu}{Seurat object / list of Seurat objects}
}
\value{
Dataframe with metadata
}
\description{
\code{extract_metadata} extracts metadata dataframe from Seurat objects
}
\examples{
data(pbmc_tiny)
tiny_list <- list(pbmc_tiny1 = pbmc_tiny, pbmc_tiny2 = pbmc_tiny)
extract_metadata(seu = tiny_list)
}
\keyword{metadata}
\keyword{preprocessing,}
\keyword{report,}
