% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{fmeasure}
\alias{fmeasure}
\title{Calculates F measure from a measures dataframe}
\usage{
fmeasure(df)
}
\arguments{
\item{df}{dataframe with measures}
}
\value{
calculated metric
}
\description{
Calculates F measure from a measures dataframe
}
\examples{
df <- data.frame(TP = 4, FP = 3, TN = 2, FN = 1)
df$ppv <- ppv(df)
df$recall <- recall(df)
fmeasure(df)
}
\keyword{metrics}
