% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{format_aberrants}
\alias{format_aberrants}
\title{Function to format DROP results
\code{format_aberrants} takes a data frame and subsets it to columns named
"sampleID", "geneID", "padjust", "type", "zScore" and "altRatio". It then
updates padjust column and takes its negative decimal logarithm, and renames
it to p_padjust (as in -log(padjust), as in pH meaning -log\link{H}).}
\usage{
format_aberrants(input_table)
}
\arguments{
\item{input_table}{A DROP results data frame or data table.}
}
\value{
A subset of the data frame containing only "sampleID", "geneID",
"p_padjust" (calculated from "padjust" column), "type", "zScore" and
"altRatio" columns, if they existed in the input data frame.
}
\description{
Function to format DROP results
\code{format_aberrants} takes a data frame and subsets it to columns named
"sampleID", "geneID", "padjust", "type", "zScore" and "altRatio". It then
updates padjust column and takes its negative decimal logarithm, and renames
it to p_padjust (as in -log(padjust), as in pH meaning -log\link{H}).
}
