% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{format_for_report}
\alias{format_for_report}
\title{Function to convert aberrant expression results into a format that allows
a heatmap report of aberrants found.
\code{format_for_report} Takes an OUTRIDER results table and reformats it,
returning only information pertaining genes aberrantly expressed in at least
one sample, and only samples containing at least one aberrantly expressed
gene. It keeps p-value and z-score information.}
\usage{
format_for_report(results, z_score_cutoff, p_adj_cutoff)
}
\arguments{
\item{z_score_cutoff}{Cutoff that gene zScore must surpass in order to be
considered aberrantly expressed.}

\item{p_adj_cutoff}{Adjusted p-value cutoff. Gene adjusted p-value must be
lower than this number in order to be considered aberrantly expressed.}
}
\value{
A table containing the merged bam stats.
}
\description{
Function to convert aberrant expression results into a format that allows
a heatmap report of aberrants found.
\code{format_for_report} Takes an OUTRIDER results table and reformats it,
returning only information pertaining genes aberrantly expressed in at least
one sample, and only samples containing at least one aberrantly expressed
gene. It keeps p-value and z-score information.
}
