% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_aberrants}
\alias{get_aberrants}
\title{Function to identify aberrantly expressed genes in an outrider results
data frame.
\code{get_aberrants} takes an outrider results data frame and creates a column
named "aberrant", set to FALSE by default. It then compares each gene's
zScore and adjusted p-value to the cutoffs provided and sets that gene's
aberrant value to TRUE if it passes the check (zScore greater than cutoff and
adjusted p-value lower than cutoff)'. It then subsets the input data frame
to samples with at least one aberrantly expressed gene and genes aberrantly
expressed in at least one sample.}
\usage{
get_aberrants(df, zScoreCutoff, padjCutoff)
}
\arguments{
\item{df}{An OUTRIDER results data frame.}

\item{zScoreCutoff}{Cutoff that gene zScore must surpass in order to be
considered aberrantly expressed.}

\item{padjCutoff}{Adjusted p-value cutoff. Gene adjusted p-value must be
lower than this number in order to be considered aberrantly expressed.}
}
\value{
A data frame containing only samples with at least one aberrantly
expressed genes and genes aberrantly expressed in at least one sample.
}
\description{
Function to identify aberrantly expressed genes in an outrider results
data frame.
\code{get_aberrants} takes an outrider results data frame and creates a column
named "aberrant", set to FALSE by default. It then compares each gene's
zScore and adjusted p-value to the cutoffs provided and sets that gene's
aberrant value to TRUE if it passes the check (zScore greater than cutoff and
adjusted p-value lower than cutoff)'. It then subsets the input data frame
to samples with at least one aberrantly expressed gene and genes aberrantly
expressed in at least one sample.
}
