% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_bam_counts}
\alias{get_bam_counts}
\title{Get counts table from rds file or table}
\usage{
get_bam_counts(file)
}
\arguments{
\item{file}{Counts table file.}
}
\value{
Loaded counts table.
}
\description{
\code{get_bam_counts} takes a file from which to extract a counts table. File may
be in RDS or table format.
}
