% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_base_ID}
\alias{get_base_ID}
\title{Get base sample ID from RNA_ID string generated for sample annotation file}
\usage{
get_base_ID(RNA_ID)
}
\arguments{
\item{RNA_ID}{A string containing the sample ID and the program used for
read counting, separated by underscores.}
}
\value{
Sample ID.
}
\description{
\code{get_base_ID} takes an RNA_ID string from the DROP sample annotation file.
String generated in this way contain not only sample ID, but the name of
the program that generated them. This function extracts the sample ID from
that string.
}
