% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_bcv}
\alias{get_bcv}
\title{Function to get the Biological Coefficient Variation of an OutriderDataSet
object.
\code{get_bcv} extracts the biological coefficient variation of an outrider
dataset.}
\usage{
get_bcv(ods)
}
\arguments{
\item{ods}{A fitted outrider dataset.}
}
\value{
A data table containing the biological coefficient variation
of the outrider dataset before and after normalisation.
}
\description{
Function to get the Biological Coefficient Variation of an OutriderDataSet
object.
\code{get_bcv} extracts the biological coefficient variation of an outrider
dataset.
}
