% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_mining.R
\name{get_cluster_string_assoc}
\alias{get_cluster_string_assoc}
\title{get_cluster_string_assoc}
\usage{
get_cluster_string_assoc(res.hcpc, string_factors)
}
\arguments{
\item{res.hcpc}{hcpc results object.}

\item{string_factors}{string factors.}
}
\value{
A data frame containing the association between each string factor
and every cluster.
}
\description{
\code{get_cluster_string_assoc} calculates the association between a cluster
and string factors.
}
\examples{
 \dontrun{
   get_cluster_string_assoc(res.hcpc, string_factors)
 }
}
