% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_counts_correlation}
\alias{get_counts_correlation}
\title{Extract correlation between samples according to gene counts.
\code{get_counts_correlation} Calculates the correlation between all samples of an
ods object according to gene count levels.}
\usage{
get_counts_correlation(ods, normalized)
}
\arguments{
\item{ods}{Outrider DataSet object.}

\item{normalized}{A boolean.
\itemize{
\item \code{TRUE} : raw counts.
\item \code{FALSE}: normalized counts.
}}
}
\value{
The correlation data frame, calculated with the spearman method.
It constains a column of clusters (calculated with euclidean distance) and a
row of EXTERNAL metadata ('yes' for imported counts and 'no' for locally
processed counts).
}
\description{
Extract correlation between samples according to gene counts.
\code{get_counts_correlation} Calculates the correlation between all samples of an
ods object according to gene count levels.
}
