% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{get_diffexp_info}
\alias{get_diffexp_info}
\title{Add a column to a counts table specifying DEG status}
\usage{
get_diffexp_info(counts_table, DEGs)
}
\arguments{
\item{counts_table}{Counts table}

\item{DEGs}{List of genes to mark as DEGs}
}
\value{
A counts table with a new column with value \code{1} for DEGs and value
\code{0} for control genes.
}
\description{
\code{get_diffexp_info} adds a column to a counts table that specifies whether
or not the gene is differentially expressed, imitating the output of counts
tables generated by the compcodeR library.
}
