% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_features_count}
\alias{get_features_count}
\title{get_features_count}
\usage{
get_features_count(results_WGCNA)
}
\arguments{
\item{results_WGCNA}{WGCNA results object.}
}
\value{
An integer. Number of traits in object.
}
\description{
\code{get_features_count} Extracts the number of traits presents in a WGCNA
results object.
}
\examples{
 \dontrun{
   get_features_count(results_WGCNA)
 }
}
