% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_gene_sample_correlations}
\alias{get_gene_sample_correlations}
\title{Get gene counts correlations across all samples of an ods object.
samples from an ods object. Genes are then clusterized.
\code{get_gene_sample_correlations} Extracts the correlation between counts across
all samples and clusterizes it.}
\usage{
get_gene_sample_correlations(
  ods,
  normalized = TRUE,
  nGenes = 500,
  rowCentered = TRUE,
  bcvQuantile = 0.9
)
}
\arguments{
\item{ods}{Outrider DataSet object.}

\item{normalized}{A boolean.
\itemize{
\item \code{TRUE} (the default) : raw counts.
\item \code{FALSE}: normalized counts.
}}

\item{nGenes}{An integer. op genes to select.}

\item{rowCentered}{A boolean.
\itemize{
\item \code{TRUE} (the default) : rows will be centered.
\item \code{FALSE}: rows will not be centered.
}}

\item{bcvQuantile}{An integer. Quantile cutoff.}
}
\value{
A data frame containing the correlations between the log2 of counts
calculated with the spearman method. It constains a column of clusters
(calculated with euclidean distance) and a row of EXTERNAL metadata ('yes'
for imported counts and 'no' for locally processed counts).
}
\description{
Get gene counts correlations across all samples of an ods object.
samples from an ods object. Genes are then clusterized.
\code{get_gene_sample_correlations} Extracts the correlation between counts across
all samples and clusterizes it.
}
