% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{get_kegg_db_path}
\alias{get_kegg_db_path}
\title{Get path for KEGG db for downloading/usage}
\usage{
get_kegg_db_path(kegg_data_file, current_organism_info, root_path)
}
\arguments{
\item{kegg_data_file}{file path. if null, inst/kegg_data_files will be used}

\item{current_organism_info}{information for the organism, including KEGG code}

\item{root_path}{only necessary if using the DEVELOPMENT mode - installs in the code dir}
}
\value{
Path to the KEGG db file - differs if in DEVEL or normal mode
}
\description{
Get path for KEGG db for downloading/usage
}
