% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_metadata}
\alias{get_metadata}
\title{Function to add rowRanges to merged counts table, and imported counts
sample information if imported counts table exists in dataset.
\code{get_metadata} takes a merged counts table, a sample annotation table,
a path to a count_ranges file and a logical, and adds some metadata to
merged counts table.}
\usage{
get_metadata(total_counts, sample_anno, count_ranges, external)
}
\arguments{
\item{total_counts}{A merged counts table.}

\item{sample_anno}{A sample annotation table.}

\item{count_ranges}{Count ranges file generated by preprocess_gtf function.}

\item{external}{A boolean.
\itemize{
\item \code{TRUE} : function will check for imported table sample information, and
update the merged_table.
\item \code{FALSE}: function will only set new rowRanges.
}}
}
\value{
A merged counts table with updated rowRanges and imported sample
metadata.
}
\description{
Function to add rowRanges to merged counts table, and imported counts
sample information if imported counts table exists in dataset.
\code{get_metadata} takes a merged counts table, a sample annotation table,
a path to a count_ranges file and a logical, and adds some metadata to
merged counts table.
}
