% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{get_org_db}
\alias{get_org_db}
\title{Obtain an org_db object from a current_organism_info, obtained from
subsetting the results of get_organism_table()}
\usage{
get_org_db(current_organism_info)
}
\arguments{
\item{current_organism_info}{row from output of get_organism_table()}
}
\value{
org_db
}
\description{
Obtain an org_db object from a current_organism_info, obtained from
subsetting the results of get_organism_table()
}
\keyword{org_db}
