% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{get_organism_table}
\alias{get_organism_table}
\title{Table with information abaut all organism available}
\usage{
get_organism_table(file = NULL)
}
\arguments{
\item{file}{to be loaded. If none given, internal organism table loaded}
}
\value{
organism table
}
\description{
Table with information abaut all organism available
}
\examples{
ot <- get_organism_table()
}
\keyword{method}
