% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{get_organismID_byOnto}
\alias{get_organismID_byOnto}
\title{Returns correct organism ID to be used}
\usage{
get_organismID_byOnto(organism_info, ont)
}
\arguments{
\item{organism_info}{organism table entry}

\item{ont}{ontology to be used}
}
\value{
organism ID to be used
}
\description{
Returns correct organism ID to be used
}
