% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_plot_df}
\alias{get_plot_df}
\title{get_plot_df}
\usage{
get_plot_df(enrich_obj, showCategory = 30)
}
\arguments{
\item{enrich_obj}{WGCNA results object.}

\item{showCategory}{Number of categories to show}
}
\value{
geneSets data.frame for printing in reports
}
\description{
\code{get_plot_df} Gets geneSets as df from enrichment object
}
\examples{
 \dontrun{
   get_plot_df(enrich_MF)
 }
}
