% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_query_distribution}
\alias{get_query_distribution}
\title{get_query_distribution}
\usage{
get_query_distribution(
  seu,
  query,
  sigfig = 3,
  layer = "scale.data",
  sample_col = "sample"
)
}
\arguments{
\item{seu}{Seurat object.}

\item{query}{Vector of query genes whose expression to analyze.}

\item{sigfig}{Significant figure cutoff.}

\item{layer}{Layer of seurat object assay to analyze.}

\item{sample_col}{Name of column that specifies sample. Default "sample",
as per our workflow.}
}
\value{
A data frame with expression levels for query genes in each sample.
}
\description{
\code{get_query_distribution} builds a table of query genes expression levels
across all samples of a Seurat object
}
\examples{
data(pbmc_tiny)
get_query_distribution(seu = pbmc_tiny, query = c("PPBP", "CA2"), sigfig = 2,
layer = "scale.data", sample_col = "orig.ident")
}
