% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_sc_markers}
\alias{get_sc_markers}
\title{get_sc_markers}
\usage{
get_sc_markers(
  seu,
  cond = NULL,
  subset_by,
  DEG = FALSE,
  logfc.threshold = 0.25,
  verbose = FALSE,
  assay = "RNA",
  values = NULL,
  min.pct = 0.1,
  layer = "scale.data"
)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{cond}{A string. Condition by which to perform DEG analysis, or by which
to group data to find conserved markers.}

\item{subset_by}{Metadata column by which seurat object will be subset for
marker calculation.}

\item{DEG}{A boolean.
\itemize{
\item \code{TRUE}: Function will calculate differentally expressed genes.
\item \code{FALSE} (the default): Function will calculate cluster markers.
}}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1
Increasing logfc.threshold speeds up the function, but can miss weaker signals.
If the \code{slot} parameter is "scale.data" no filtering is performed.}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}

\item{assay}{A string. Assay whose markers will be calculated. Default "RNA",
as per usual workflow.}

\item{values}{A comma-separated string delimiting values of cond column by
which to perform the comparison. Must specify two conditions. Default NULL,
will use all values. If there are more than two, function will return an
error.}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{layer}{Name of layer to get or set}
}
\value{
A list containing one marker or DEG data frame per cluster, plus
an additional one for global DEGs if performing differential analysis.
}
\description{
\code{get_sc_markers} performs differential expression analysis on OR selects
conserved markers from a seurat object.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- 1
pbmc_tiny$seurat_clusters[8:15] <- 2
get_sc_markers(seu = pbmc_tiny, cond = "groups", DEG = TRUE, verbose = TRUE,
               subset_by = "seurat_clusters")
}
