% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{get_stats}
\alias{get_stats}
\title{Function to calculate full set of measures from a confusion data frame}
\usage{
get_stats(df)
}
\arguments{
\item{df}{confusion dataframe}
}
\value{
a data frame with several measures
}
\description{
Function to calculate full set of measures from a confusion data frame
}
\examples{
df <- data.frame(TP = 4, FP = 3, TN = 2, FN = 1)
stats <- get_stats(df)
}
