% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{get_stats_from_cm}
\alias{get_stats_from_cm}
\title{Obtain full set of measures from a confusion matrix}
\usage{
get_stats_from_cm(cm)
}
\arguments{
\item{cm}{confusion matrix}
}
\value{
a data frame with several measures
}
\description{
Obtain full set of measures from a confusion matrix
}
\examples{
\dontrun{
cm <- matrix(c(4,3,2,1), ncol = 2)
rownames(cm) <- c(TRUE,FALSE)
colnames(cm) <- c(TRUE,FALSE)
stats <- get_stats_from_cm(cm)
}
}
