% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_unique_rownames}
\alias{get_unique_rownames}
\title{Get unique rownames of a data frame}
\usage{
get_unique_rownames(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
Unique rownames of data frame.
}
\description{
\code{get_unique_rownames} takes a data frame and extracts its unique row names.
}
