% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{gg_heatmap}
\alias{gg_heatmap}
\title{gg_heatmap}
\usage{
gg_heatmap(
  data_table,
  input = "",
  transpose = FALSE,
  y_axis = "y_axis",
  x_axis = "x_axis",
  fill = "Freq",
  text_plot = NULL,
  labs = TRUE,
  x_angle = 25,
  text_size = 2.5,
  text_colour = "black",
  col = c("#0000D5", "#FFFFFF", "#D50000"),
  na_col = "grey50"
)
}
\arguments{
\item{data_table}{Input data table.}

\item{input}{Input string. Only used if it is equal to "matrix", in which
case it will convert input data to a data frame.}

\item{transpose}{A boolean.
\itemize{
\item \code{TRUE}: Data will be transposed.
\item \code{FALSE} (the default): Data will NOT be transposed.
}}

\item{x_axis, y_axis}{Names of X and Y axes.}

\item{fill}{Fill colour. \code{fill_alpha()} can be used to set the transparency
of the fill.}

\item{text_plot}{Label to add to plot.}

\item{labs}{A boolean.
\itemize{
\item \code{TRUE} (the default): Data will be transposed.
\item \code{FALSE}: Data will NOT be transposed.
}}

\item{x_angle}{Angle by which x axis labels will be rotated.}

\item{text_size, text_colour}{Font size and colour for text in plot}

\item{col}{Color pallete to use when plotting.}

\item{na_col}{Color for NA values.}
}
\value{
A ggplot2 heatmap.
}
\description{
\code{gg_heatmap} Builds a ggplot2 heatmap from a data table.
}
\examples{
 \dontrun{
   gg_heatmap(matrix, input = "matrix")
 }
}
