% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{\%:::\%}
\alias{\%:::\%}
\title{Method used to import hidden functions legally for CRAN check}
\usage{
pkg \%:::\% fun
}
\arguments{
\item{pkg}{package}

\item{fun}{function to be luaded}
}
\value{
function requried
}
\description{
Method used to import hidden functions legally for CRAN check
}
