% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{graph_file_name}
\alias{graph_file_name}
\title{Generates the full path to a specific graph file}
\usage{
graph_file_name(path, file_name, ext = "pdf")
}
\arguments{
\item{path}{to file}

\item{file_name}{file name}

\item{ext}{file extension}
}
\value{
full file path and name
}
\description{
Generates the full path to a specific graph file
}
\keyword{file}
