% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{id_translation_orgdb}
\alias{id_translation_orgdb}
\title{Translates a given gene ID}
\usage{
id_translation_orgdb(input_ids, organism_db, org_var_name)
}
\arguments{
\item{input_ids}{gene IDs to be translated}

\item{organism_db}{BiocGenerics organism database}

\item{org_var_name}{BiocGenerics organism translation variable name}
}
\value{
translated gene IDs
}
\description{
Translates a given gene ID
}
\keyword{translate}
