% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installFunc.R
\name{install_DEgenes_hunter}
\alias{install_DEgenes_hunter}
\title{Hunter installation Function}
\usage{
install_DEgenes_hunter(to)
}
\arguments{
\item{to}{Path to folder in which copy Hunter scripts}
}
\value{
void
}
\description{
This function allows you to install scripts.
}
\examples{
install_DEgenes_hunter("./")
}
\keyword{installation}
