% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{load_DEG_output}
\alias{load_DEG_output}
\title{load output from previous execution of DEG module}
\usage{
load_DEG_output(
  targets,
  load_path,
  min_avg_log2FC,
  min_cell_proportion,
  p_val_cutoff
)
}
\arguments{
\item{targets}{Vector of target names.}

\item{load_path}{Path where target executions will be found.}

\item{min_avg_log2FC}{Average log2fc cutoff for significant DEGs.
Default 0.5.}

\item{min_cell_proportion}{Minimum threshold of percentage of cells
expressing DEG. Despite Seurat::FindMarkers being called "min.pct", it is
not a percentage but a value from 0 to 1.}

\item{p_val_cutoff}{Adjusted p-val cutoff for significant DEGs.
Default 5e-3.}
}
\value{
A DEG results list with identical structure to main_sc_Hunter
output.
}
\description{
\code{load_DEG_output} loads a previous execution of our Single-Cell DEA.
}
\examples{
\dontrun{
   load_DEG_output(targets = c("42_vs_68", "Ctrl_vs_Mut"),
                   load_path = "path/to/results")
}
}
