% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{load_hunter_folder}
\alias{load_hunter_folder}
\title{Function to load the expression hunter folder
Used by the functional_Hunter script}
\usage{
load_hunter_folder(path = NULL)
}
\arguments{
\item{path}{where the expression hunter results are}
}
\value{
object containing expression, DE and WGCNA data
}
\description{
Function to load the expression hunter folder
Used by the functional_Hunter script
}
\examples{
# Can only be run after running the write_expression_data function 
# eg as part of the degenes_Hunter.R script
\dontrun{
hunter_results <- load_hunter_folder(input_hunter_folder)
}
}
