% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{load_package_result}
\alias{load_package_result}
\title{Loads stored results in DEgenes Hunter expression analysis package folder and returns in compact object}
\usage{
load_package_result(pack_path)
}
\arguments{
\item{pack_path}{of expression package folder}
}
\value{
package info laoded
}
\description{
Loads stored results in DEgenes Hunter expression analysis package folder and returns in compact object
}
\keyword{input}
