% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{load_synth_dataset}
\alias{load_synth_dataset}
\title{Loads Hunter DEG analysis table and simulated TRUE DEG vector. Concat in
training/testing ML data frame format and returns it.}
\usage{
load_synth_dataset(
  folder = NULL,
  hunter_table_file = "hunter_results_table.txt",
  prediction_file_regex = "_predv"
)
}
\arguments{
\item{folder}{where files are placed}

\item{hunter_table_file}{DEG analys result table file name}

\item{prediction_file_regex}{regex to find TRUE DEG file generated}
}
\value{
a data frame with DEG result and real predictions column
}
\description{
Loads Hunter DEG analysis table and simulated TRUE DEG vector. Concat in
training/testing ML data frame format and returns it.
}
\examples{
# Used to load a synth + DE analysis result folder
ht <- load_synth_dataset() 
# Will return an empty object, you need a correct folder
}
