% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{load_WGCNA_results}
\alias{load_WGCNA_results}
\title{Loads stored WGCNA results in DEgenes Hunter expression analysis package folder and returns in compact object}
\usage{
load_WGCNA_results(path, main_deg_table)
}
\arguments{
\item{path}{of expression analysis results (Hunter folder)}

\item{main_deg_table}{dataframe with main DEG analysis results}
}
\value{
WGCNA results loaded
}
\description{
Loads stored WGCNA results in DEgenes Hunter expression analysis package folder and returns in compact object
}
\keyword{input}
