% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_aberrant_expression.R
\name{main_aberrant_expression}
\alias{main_aberrant_expression}
\title{Main DROP Aberrant Expression function}
\usage{
main_aberrant_expression(
  sample_annotation = NULL,
  anno_database = NULL,
  count_ranges = NULL,
  gene_mapping_file = NULL,
  count_files = NULL,
  dataset = NULL,
  cpu = 1,
  config_file = NULL,
  hpo_file = NULL,
  sample_bam_stats = NULL,
  top_N = 10
)
}
\arguments{
\item{sample_annotation}{Path to sample annotation table. See documentation.}

\item{anno_database}{TxDb annotation database created by preprocess_gtf
function.}

\item{count_ranges}{Count ranges file generated by preprocess_gtf function.}

\item{gene_mapping_file}{Gene mapping file generated by preprocess_gtf
function.}

\item{count_files}{Count tables to process. Can be in rds or table format.}

\item{dataset}{Name of the dataset.}

\item{cpu}{Amount of CPUs available for analysis.}

\item{config_file}{Configuration file in yaml format.}

\item{hpo_file}{Genes and associated HPO terms table. Optional.}

\item{sample_bam_stats}{Bam stats of each sample.Generated with samtools
idxstats (see Aberrant_Expression workflow for how to generate this file).}

\item{top_N}{Top N genes by adjusted p-value to select for OUTRIDER
overview. Default: 10.}
}
\value{
Aberrant expression object containing all analysis results.
}
\description{
\code{main_aberrant_expression} runs DROP OUTRIDER analysis.
}
