% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_abgenes_Hunter.R
\name{main_abgenes_Hunter}
\alias{main_abgenes_Hunter}
\title{Main Aberrant Expression function}
\usage{
main_abgenes_Hunter(
  sample_annotation = NULL,
  anno_database = NULL,
  count_ranges = NULL,
  gene_mapping_file = NULL,
  count_files = NULL,
  dataset = NULL,
  cpu = 1,
  fpkm_cutoff = 1,
  implementation = "autoencoder",
  max_dim_proportion = 3,
  p_adj_cutoff = 0.05,
  z_score_cutoff = 0.2,
  hpo_file = NULL,
  stats_path = NULL,
  top_N = 10
)
}
\arguments{
\item{sample_annotation}{Path to sample annotation table. See documentation.}

\item{anno_database}{TxDb annotation database created by preprocess_gtf
function.}

\item{count_ranges}{Count ranges file generated by preprocess_gtf function.}

\item{gene_mapping_file}{Gene mapping file generated by preprocess_gtf
function.}

\item{count_files}{Count tables to process. Can be in rds or table format.}

\item{dataset}{Name of the dataset.}

\item{cpu}{Amount of CPUs available for analysis. Default 1.}

\item{fpkm_cutoff}{Min FPKM value for read filtering. Default 3.}

\item{implementation}{Method for sample covariation removal in OUTRIDER.
Possible values: "autoencoder" (default), "pca" or "peer".}

\item{max_dim_proportion}{Maximum value for autoencoder encoding
dimension. Default 3, optimum for aberrant expression.}

\item{p_adj_cutoff}{Adjusted p-value cutoff. Gene adjusted p-value must be
lower than this number in order to be considered aberrantly expressed.}

\item{z_score_cutoff}{Cutoff that gene zScore must surpass in order to be
considered aberrantly expressed.}

\item{hpo_file}{File containing HPO terms associated to genes. Optional.}

\item{stats_path}{Bam stats directories of each sample. Stats generated
with samtools idxstats (see Aberrant_Expression workflow for how to generate
this file).}

\item{top_N}{Top N genes by adjusted p-value to select for OUTRIDER
overview. Default 10.}
}
\value{
Aberrant expression object containing all analysis results.
}
\description{
\code{main_abgenes_Hunter} runs DROP OUTRIDER analysis.
}
