% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_degsynth.R
\name{main_compcodeR}
\alias{main_compcodeR}
\title{main_compcodeR}
\usage{
main_compcodeR(
  dataset,
  n.vars,
  samples.per.cond,
  n.diffexp,
  repl.id = 1,
  seqdepth = 1e+07,
  minfact = 0.7,
  maxfact = 1.4,
  relmeans = "auto",
  dispersions = "auto",
  fraction.upregulated = 1,
  between.group.diffdisp = FALSE,
  filter.threshold.total = 1,
  filter.threshold.mediancpm = 0,
  fraction.non.overdispersed = 0,
  random.outlier.high.prob = 0,
  random.outlier.low.prob = 0,
  single.outlier.high.prob = 0,
  single.outlier.low.prob = 0,
  effect_sizes = 1.5,
  tree = NULL,
  prop.var.tree = 1,
  nDEGs = as.character(n.vars/2),
  model.process = c("BM", "OU"),
  selection.strength = 0,
  id.condition = NULL,
  overlap_size = 0,
  id.species = as.factor(rep(1, 2 * samples.per.cond)),
  check.id.species = TRUE,
  lengths.relmeans = NULL,
  lengths.dispersions = NULL,
  lengths.phylo = TRUE,
  output_dir = NULL,
  method = "vanilla",
  fixed_DEG_lists = "",
  fixed_upregulated_DEGs = "",
  condition_columns,
  exp_design = "",
  inputfile = NULL
)
}
\arguments{
\item{dataset}{A name or identifier for the data set/simulation settings.}

\item{n.vars}{The initial number of genes in the simulated data set. Based on the filtering conditions (\code{filter.threshold.total} and \code{filter.threshold.mediancpm}), the number of genes in the final data set may be lower than this number.}

\item{samples.per.cond}{The number of samples in each of the two conditions.}

\item{n.diffexp}{The number of genes simulated to be differentially expressed between the two conditions.}

\item{repl.id}{A replicate ID for the specific simulation instance. Useful for example when generating multiple count matrices with the same simulation settings.}

\item{seqdepth}{The base sequencing depth (total number of mapped reads). This number is multiplied by a value drawn uniformly between \code{minfact} and \code{maxfact} for each sample to generate data with different actual sequencing depths.}

\item{minfact, maxfact}{The minimum and maximum for the uniform distribution used to generate factors that are multiplied with \code{seqdepth} to generate individual sequencing depths for the simulated samples.}

\item{relmeans}{A vector of mean values to use in the simulation of data from the Negative Binomial distribution, or \code{"auto"}. Note that these values may be scaled in order to comply with the given sequencing depth. With the default value (\code{"auto"}), the mean values are sampled from values estimated from the Pickrell and Cheung data sets. If \code{relmeans} is a vector, the provided values will be used as mean values in the simulation for the samples in the first condition. The mean values for the samples in the second condition are generated by combining the \code{relmeans} and \code{effect.size} arguments.}

\item{dispersions}{A vector or matrix of dispersions to use in the simulation of data from the Negative Binomial distribution, or \code{"auto"}. With the default value (\code{"auto"}), the dispersion values are sampled from values estimated from the Pickrell and Cheung data sets. If both \code{relmeans} and \code{dispersions} are set to \code{"auto"}, the means and dispersion values are sampled in pairs from the values in these data sets. If \code{dispersions} is a single vector, the provided dispersions will be used for simulating data from both conditions. If it is a matrix with two columns, the values in the first column are used for condition 1, and the values in the second column are used for condition 2.}

\item{fraction.upregulated}{The fraction of the differentially expressed genes that is upregulated in condition 2 compared to condition 1.}

\item{between.group.diffdisp}{Whether or not the dispersion should be allowed to be different between the conditions. Only applicable if \code{dispersions} is \code{"auto"}.}

\item{filter.threshold.total}{The filter threshold on the total count for a gene across all samples. All genes for which the total count across all samples is less than the threshold will be filtered out.}

\item{filter.threshold.mediancpm}{The filter threshold on the median count per million (cpm) for a gene across all samples. All genes for which the median cpm across all samples is less than the threshold will be filtered out.}

\item{fraction.non.overdispersed}{The fraction of the genes that should be simulated according to a Poisson distribution, without overdispersion. The non-overdispersed genes will be divided proportionally between the upregulated, downregulated and non-differentially expressed genes.}

\item{random.outlier.high.prob}{The fraction of 'random' outliers with unusually high counts.}

\item{random.outlier.low.prob}{The fraction of 'random' outliers with unusually low counts.}

\item{single.outlier.high.prob}{The fraction of 'single' outliers with unusually high counts.}

\item{single.outlier.low.prob}{The fraction of 'single' outliers with unusually low counts.}

\item{tree}{a dated phylogenetic tree of class \code{\link[ape]{phylo}} with `samples.per.cond * 2` species.}

\item{prop.var.tree}{the proportion of the common variance explained by the tree for each gene. It can be a scalar, in which case the same parameter is used for all genes. Otherwise it needs to be a vector with length \code{n.vars}. Default to 1.}

\item{model.process}{the process to be used for phylogenetic simulations. One of "BM" or "OU", default to "BM".}

\item{selection.strength}{if the process is "OU", the selection strength parameter.}

\item{id.condition}{A named vector, indicating which species is in each condition. Default to first `samples.per.cond` species in condition `1` and others in condition `2`.}

\item{id.species}{A factor giving the species for each sample. If a tree is used, should be a named vector with names matching the taxa of the tree. Default to \code{rep(1, 2*samples.per.cond)}, i.e. all the samples come from the same species.}

\item{check.id.species}{Should the species vector be checked against the tree lengths (if provided) ? If TRUE, the function checks that all the samples that share a factor value in \code{id.species} that their distance on the tree is zero, i.e. that they are on the same tip of the tree. Default to TRUE.}

\item{lengths.relmeans}{An optional vector of mean values to use in the simulation of lengths from the Negative Binomial distribution. Should be of length n.vars. Default to \code{NULL}: the lengths are not taken into account for the simulation. If set to \code{"auto"}, the mean length values are sampled from values estimated from the Stern & Crandall (2018) data set.}

\item{lengths.dispersions}{An optional vector of dispersions to use in the simulation of data from the Negative Binomial distribution. Should be of length n.vars. Default to \code{NULL}: the lengths are not taken into account for the simulation. If set to \code{"auto"}, the dispersion length values are sampled from values estimated from the Stern & Crandall (2018) data set.}

\item{lengths.phylo}{If TRUE, the lengths are simulated according to a phylogenetic Poisson Log-Normal model on the tree, with a BM process. If FALSE, they are simulated according to an iid negative binomial distribution. In both cases, \code{lengths.relmeans} and \code{lengths.dispersions} are used. Default to TRUE if a tree is provided.}

\item{output_dir}{Directory where counts matrix and DEG table will be saved.}
}
\value{
A list. Element \code{all} contains the entire results object. Element
\code{counts_table} contains just the counts matrix. Element \code{DEGs} contains
a data frame detailing which genes are differentially expressed in dataset.
}
\description{
\code{main_compcodeR} Is a wrapper for the \code{generateSyntheticData} function
from package \code{compcodeR}.
}
\examples{
B_625_625 <- main_compcodeR(dataset = "B_625_625", n.vars = 12500, 
                                 samples.per.cond = 5, n.diffexp = 1250, 
                                 repl.id = 1, seqdepth = 1e7, 
                                 fraction.upregulated = 0.5, 
                                 between.group.diffdisp = FALSE, 
                                 filter.threshold.total = 1, 
                                 filter.threshold.mediancpm = 0, 
                                 fraction.non.overdispersed = 0)
B_625_625
}
