% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{make_DEG_table}
\alias{make_DEG_table}
\title{Create synthetic DEG table
\code{make_DEG_table} adds a synthetic differential expression effect to a
counts table. It controls effect size, number of samples, standard deviation
and condition to consider.}
\usage{
make_DEG_table(
  counts_table,
  exp_design,
  factor_column,
  up_DEGs,
  effect_size,
  deg_method,
  DEGs,
  samples_per_group,
  stdev = 0.2
)
}
\arguments{
\item{counts_table}{Counts table to modify}

\item{exp_design}{Experimental design}

\item{factor_column}{Index of column whose factor to consider}

\item{up_DEGs}{Genes to upregulate}

\item{effect_size}{Effect size to emulate in expression values}

\item{deg_method}{Method to use. Values: "norm" (\code{synth_normal_DEGs}), det
(\code{synth_deterministic_DEGs}) and "exp" (\code{synth_exponential_DEGs})}

\item{DEGs}{List of all differentially expressed genes}

\item{samples_per_group}{Number of samples that each group should have}

\item{stdev}{Standard deviation for distribution. Used in \code{synth_normal_DEGs}}
}
\value{
Table with differential expression emulation
}
\description{
Create synthetic DEG table
\code{make_DEG_table} adds a synthetic differential expression effect to a
counts table. It controls effect size, number of samples, standard deviation
and condition to consider.
}
