% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{mark_duplicates}
\alias{mark_duplicates}
\title{Add tag to duplicate values in vector
\code{mark_duplicates} finds duplicate values inside a vector and appends
a tag to each of them. This tag can either be an index or a letter}
\usage{
mark_duplicates(string, index_as_letter = FALSE)
}
\arguments{
\item{string}{A string}

\item{index_as_letter}{A boolean
\itemize{
\item \code{TRUE}:  Duplicates will be marked as value_A, value_B
\item \code{FALSE} (the default): Duplicates will be marked as value_1, value_2
}}
}
\value{
A vector with marked duplicates
}
\description{
Add tag to duplicate values in vector
\code{mark_duplicates} finds duplicate values inside a vector and appends
a tag to each of them. This tag can either be an index or a letter
}
