% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{match_cell_types}
\alias{match_cell_types}
\title{match_cell_types
takes a cluster-marker gene data frame and a cell type
marker file. It then looks for matches between the two and assigns a cell
type to each cluster of the data frame.}
\usage{
match_cell_types(markers_df, cell_annotation, p_adj_cutoff = 1e-05)
}
\arguments{
\item{markers_df}{Data frame of markers, clusters and p-values}

\item{cell_annotation}{Table of cell types and their associated markers}

\item{p_adj_cutoff}{Minimum adjusted p-value of markers to consider.}
}
\value{
A markers data frame with a new column for cell type assigned to
cluster.
}
\description{
match_cell_types
takes a cluster-marker gene data frame and a cell type
marker file. It then looks for matches between the two and assigns a cell
type to each cluster of the data frame.
}
\examples{
 \dontrun{
   match_cell_types(markers_df = markers_df, p_adj_cutoff = 1e-5,
                    cell_annotation = markers_celltypes_df)
 }
}
