% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{match_counts_to_design}
\alias{match_counts_to_design}
\title{Sort counts table by samples in experimental design
\code{match_counts_to_design} makes sure that samples are sorted in the same order
in counts table and experimental design.}
\usage{
match_counts_to_design(counts_table, exp_design)
}
\arguments{
\item{counts_table}{Counts table}

\item{exp_design}{Experimental design}
}
\value{
Reordered counts table
}
\description{
Sort counts table by samples in experimental design
\code{match_counts_to_design} makes sure that samples are sorted in the same order
in counts table and experimental design.
}
