% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{merge_counts}
\alias{merge_counts}
\title{Function to extract count tables and merge them into one.
\code{merge_counts} takes multiple paths to count tables and reads them, merging
them into one.}
\usage{
merge_counts(cpu, sample_anno, count_files, count_ranges)
}
\arguments{
\item{cpu}{Amount of CPUs available for analysis. Default 1.}

\item{sample_anno}{A sample annotation table.}

\item{count_files}{Count tables to process. Can be in rds or table format.}

\item{count_ranges}{Count ranges file generated by preprocess_gtf function.}
}
\value{
A merged counts table.
}
\description{
Function to extract count tables and merge them into one.
\code{merge_counts} takes multiple paths to count tables and reads them, merging
them into one.
}
