% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{merge_seurat}
\alias{merge_seurat}
\title{merge_seurat
\code{merge_seurat} loads single-cell count matrices and creates a merged
seurat object.}
\usage{
merge_seurat(project_name, exp_design, count_path, suffix = "")
}
\arguments{
\item{project_name}{Will appear in output project.name slot.}

\item{exp_design}{Experiment design table in TSV format.}

\item{count_path}{Directory that will be searched for count matrices.}

\item{suffix}{Suffix to paste with main count_path. Avoids unnecessary
globbing.}
}
\value{
Merged Seurat object
}
\description{
merge_seurat
\code{merge_seurat} loads single-cell count matrices and creates a merged
seurat object.
}
\examples{
 \dontrun{
   merge_seurat(project_name = "experiment_name", exp_design = exp_design,
                count_path = "path/to/counts/folder", suffix = "path/suffix")
 }
}
\keyword{integration}
\keyword{merging,}
\keyword{preprocessing,}
