% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{multienricher}
\alias{multienricher}
\title{Perform enrichment of ORA and/or GSEA using a set of genes given}
\usage{
multienricher(
  genes,
  organism_info,
  keytype = "ENTREZID",
  ontology = "bkr",
  enrichmentType = "o",
  pvalueCutoff = 0.01,
  pAdjustMethod = "BH",
  useInternal = TRUE,
  qvalueCutoff = 0.02,
  cores = 1,
  task_size = 1,
  verbose = FALSE
)
}
\arguments{
\item{genes}{significant genes: vector or list of vectors for each cluster}

\item{organism_info}{organism table info. Must include entries depending on
ontologies selected: Bioconductor_DB (GO), KeggCode (KEGG), Reactome_ID
(Reactome)}

\item{keytype}{gene code type. If is a vector, will be used in the same
order than ontology items are given}

\item{ontology}{string with ontologIDs to be used. Allowed: Gene Ontology
(BP: b ; MF: m ; CC: c), KEGG (k), Reactome (r)}

\item{enrichmentType}{enrichment technique to be used. Allowed: ORA (o) and
GSEA (g)}

\item{pvalueCutoff}{p-value threshold}

\item{pAdjustMethod}{p-value adjust method to be applied}

\item{useInternal}{optional KEGG internal usage flag}

\item{qvalueCutoff}{q-value threshold. ONLY USED FOR ORA.}

\item{cores}{for parallel. Only used in case of genes are cluster genes}

\item{task_size}{number of elements per packages used}

\item{verbose}{activate verbose mode}
}
\value{
list with enrichments performed
}
\description{
Perform enrichment of ORA and/or GSEA using a set of genes given
}
\examples{
data(degh_output)
genes <- head(rownames(degh_output$raw_filter),1000) 
# You need to translate to entrez if you want use all ontology types
ontologies <- "" # Select your wanted ontologies
organisms_table <- get_organism_table()
current_organism_info <- subset(organisms_table, 
                         rownames(organisms_table) == "Human")
enrch <- multienricher(genes = genes, organism_info = current_organism_info,
ontology = "") # Return NULL.
}
\keyword{enrich}
