% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library_new.R
\name{multienricher_topGO}
\alias{multienricher_topGO}
\title{Perform topGO enrichment analysis of a list of genes}
\usage{
multienricher_topGO(
  all_funsys,
  genes_list,
  universe = NULL,
  organism_info,
  gene_id = "entrez",
  ...
)
}
\description{
Perform topGO enrichment analysis of a list of genes
}
