% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{obtain_info_from_biomaRt}
\alias{obtain_info_from_biomaRt}
\title{Obtains biological info from BiomaRt API}
\usage{
obtain_info_from_biomaRt(orthologues, id_type, mart, dataset, host, attr)
}
\arguments{
\item{orthologues}{orthologue organism ID}

\item{id_type}{gene code type}

\item{mart}{BioMart database name you want to connect to}

\item{dataset}{Dataset you want to use}

\item{host}{host to connect to}

\item{attr}{Attributes you want to retrieve}
}
\value{
downloaded biomaRt response
}
\description{
Obtains biological info from BiomaRt API
}
\keyword{method}
