% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{parallel_list}
\alias{parallel_list}
\title{Parallelize function execution in list with multiple elements}
\usage{
parallel_list(X, FUNC, workers = 2, task_size = 1, ...)
}
\arguments{
\item{X}{R list}

\item{FUNC}{Function to execute in each element of the list}

\item{workers}{Number of process to use in parallel execution}

\item{task_size}{Number of list elements to be executed on each
BiocParallel task}

\item{...}{Additional arguments to BiocParallel or to FUNC}
}
\value{
list
}
\description{
Parallelize function execution in list with multiple elements
}
\keyword{performance}
