% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{parse_filter}
\alias{parse_filter}
\title{parse_filter}
\usage{
parse_filter(object, expression)
}
\arguments{
\item{object}{Object to filter.}

\item{expression}{Expression to parse as filter.}
}
\value{
filtered object
}
\description{
\code{parse_filter} parses an object and a string into an expression that filters the object.
}
\examples{
\dontrun{
     data_frame <- head(mtcars)
     filter <- parse_filter(object = "data_frame", expression = "mpg > 20")
     data_frame[filter, ]
  }
}
