% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_tiny}
\alias{pbmc_tiny}
\title{Subset of the pbmc dataset distributed in package SeuratObject}
\format{
A seurat object containing the 15 first samples of the
pbmc_small dataset.
}
\source{
Created with data-raw/pbmc_tiny.R. Original source:
https://search.r-project.org/CRAN/refmans/SeuratObject/html/pbmc_small.html
}
\usage{
data(pbmc_tiny)
}
\description{
Subset of the pbmc dataset distributed in package SeuratObject
}
\keyword{datasets}
