% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{perform_GSEA_clusters}
\alias{perform_GSEA_clusters}
\title{Performs GSEA enrichment over several gene sets (clusters)}
\usage{
perform_GSEA_clusters(
  all_clusters,
  organism,
  keyType,
  pvalueCutoff,
  pAdjustMethod = "BH",
  ont,
  useInternal = FALSE
)
}
\arguments{
\item{all_clusters}{list of gene sets (clusters)}

\item{organism}{target organism}

\item{keyType}{gene code type}

\item{pvalueCutoff}{p-value threshold}

\item{pAdjustMethod}{p-value adjust method to be applied}

\item{ont}{ontology to be used}

\item{useInternal}{optional KEGG param used to indicate if already
downloaded KEGG database must be used or online API must be called.
DEfault: FALSE}
}
\value{
enrichment tables obtained
}
\description{
Performs GSEA enrichment over several gene sets (clusters)
}
\keyword{enrich}
