% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{perform_topGO}
\alias{perform_topGO}
\title{Run topGO enrichment and generates target plots storing into PDF files}
\usage{
perform_topGO(
  attr_name,
  interesting_genenames,
  DEG_annot_table,
  ontology,
  graphname,
  filter_name,
  output_files
)
}
\arguments{
\item{attr_name}{target column of DEG_annot_table with attributes}

\item{interesting_genenames}{target genes}

\item{DEG_annot_table}{DEgenes Hunter expression result table}

\item{ontology}{GO subontology to be used}

\item{graphname}{output file name}

\item{filter_name}{target column of DEG_annot_table with filtering param}

\item{output_files}{output path}
}
\value{
void
}
\description{
Run topGO enrichment and generates target plots storing into PDF files
}
\keyword{enrich}
