% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{perform_topGO_local}
\alias{perform_topGO_local}
\title{Pipeline to perform topGO enrichment}
\usage{
perform_topGO_local(
  entrez_targets,
  entrez_universe,
  sub_ontology,
  outFile = NULL,
  organism,
  plot_graph = TRUE
)
}
\arguments{
\item{entrez_targets}{genes used to enrich}

\item{entrez_universe}{total of allowed genes}

\item{sub_ontology}{GO subontology to be used}

\item{outFile}{output path}

\item{organism}{target organism}

\item{plot_graph}{boolean to generates a plot}
}
\value{
enrichment tables obtained
}
\description{
Pipeline to perform topGO enrichment
}
\keyword{enrich}
