% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{plotAllelicCounts}
\alias{plotAllelicCounts}
\title{plotAllelicCounts}
\usage{
plotAllelicCounts(
  data,
  title = NULL,
  padjCutoff = 0.05,
  allelicRatioCutoff = 0.8,
  rare_column = NULL
)
}
\arguments{
\item{data}{A data.frame containing the counts or results table from \code{DESeq4MAE} function}

\item{title}{The plot's title}

\item{padjCutoff}{The significance level}

\item{allelicRatioCutoff}{The minimum allelic ratio ALT/(ALT+REF) to be
considered signficant}

\item{rare_column}{The name of the column that indicates if a variant is rare or not.
Default is \code{null} which means it won't be plotted.}
}
\value{
A ggplot object containing the MA plot.
}
\description{
Creates an allelic counts plot, ie, counts of the alternative
vs counts of the reference on the log10 scale. If significant and minor
allele frequency columns present, dots will be highlighted accordingly.
}
\examples{
file <- system.file("extdata", "allelic_counts_HG00187.csv", 
    package = "tMAE", mustWork = TRUE)
maeCounts <- fread(file)
res <- DESeq4MAE(maeCounts)
plotAllelicCounts(res)

}
\author{
Vicente Yepez
}
