% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{plotMA4MAE}
\alias{plotMA4MAE}
\title{plotMA4MAE}
\usage{
plotMA4MAE(
  data,
  title = NULL,
  padjCutoff = 0.05,
  allelicRatioCutoff = 0.8,
  rare_column = NULL
)
}
\arguments{
\item{data}{A data.frame containing the results table from \code{DESeq4MAE} function}

\item{title}{The plot's title}

\item{padjCutoff}{The significance level}

\item{allelicRatioCutoff}{The minimum allelic ratio ALT/(ALT+REF) to be
considered signficant}

\item{rare_column}{The name of the column that indicates if a variant is rare or not.
Default is \code{null} which means it won't be plotted.}
}
\value{
A ggplot object containing the MA plot.
}
\description{
Creates an MA plot, ie, Fold Change (ALT/REF) vs Coverage colored
by significance (from p adjusted and allelic ratio) or significance and
minor allele frequency (if rare_column is provided).
}
\examples{
file <- system.file("extdata", "allelic_counts_HG00187.csv",
    package = "tMAE", mustWork = TRUE)
maeCounts <- fread(file)
res <- DESeq4MAE(maeCounts)
plotMA4MAE(res)
}
\author{
Vicente Yepez
}
