% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{ppv}
\alias{ppv}
\title{Calculates Positive Predictive Value from a measures dataframe}
\usage{
ppv(df)
}
\arguments{
\item{df}{dataframe with measures}
}
\value{
calculated metric
}
\description{
Calculates Positive Predictive Value from a measures dataframe
}
\examples{
df <- data.frame(TP = 4, FP = 3, TN = 2, FN = 1)
df$Precision <- ppv(df)
}
\keyword{metrics}
