% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{prepare_for_fortify}
\alias{prepare_for_fortify}
\title{Mutate enrichment object to remove conflict of IDs or Descriptions for
fortify}
\usage{
prepare_for_fortify(enr)
}
\arguments{
\item{enr}{enrichment object}
}
\value{
enrichment object with repeated IDs or Description cleaned
}
\description{
Mutate enrichment object to remove conflict of IDs or Descriptions for
fortify
}
