% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{preprocess_gtf}
\alias{preprocess_gtf}
\title{Preprocess GTF for DROP analysis}
\usage{
preprocess_gtf(gtf)
}
\arguments{
\item{gtf}{GTF file to process.}
}
\value{
A list. Items txdb contains the database, count_ranges describes
known exons in the genome, gene_name_mapping condenses gene information.
gene_name_mapping contains several columns: chromosome, nucleotide start,
nucleotide end, strand, gene ID, gene name, gene type and original gene name.
}
\description{
\code{preprocess_gtf} takes a GTF file and a directory path. It outputs a txdb,
count ranges, and gene_mapping files, which it writes to provided directory.
}
\examples{
gtf <- system.file("extData/testdata", "gencode.v45.toy.annotation.gtf",
                    package = "ExpHunterSuite")
processed <- preprocess_gtf(gtf)
processed$txdb
head(processed$count_ranges)
head(processed$gene_name_mapping)
}
