% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{processed_vs_imported}
\alias{processed_vs_imported}
\title{Function to separate locally processed and GTEx samples from OUTRIDER
results.
\code{processed_vs_imported} takes a data frame and splits it into two data frames
contained in a list. The two elements are "processed" and "imported".
All samples whose name contain "GTEX" are considered imported, the rest
are considered exported.}
\usage{
processed_vs_imported(df)
}
\arguments{
\item{df}{A data frame containing read counts by sample.}
}
\value{
A list with two elements, one for processed counts and another
for imported counts.
}
\description{
Function to separate locally processed and GTEx samples from OUTRIDER
results.
\code{processed_vs_imported} takes a data frame and splits it into two data frames
contained in a list. The two elements are "processed" and "imported".
All samples whose name contain "GTEX" are considered imported, the rest
are considered exported.
}
