% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{read_DEG_lists}
\alias{read_DEG_lists}
\title{Read list of DEGs from disk
\code{read_DEG_lists} was developed for testing purposes. It loads a list of DEGs
from disk and returns it in a format that makes it usable in our DEG
synthesis library.}
\usage{
read_DEG_lists(paths_string)
}
\arguments{
\item{paths_string}{Comma-separated paths from which to read DEG lists}
}
\value{
A list containing all DEG lists.
}
\description{
Read list of DEGs from disk
\code{read_DEG_lists} was developed for testing purposes. It loads a list of DEGs
from disk and returns it in a format that makes it usable in our DEG
synthesis library.
}
