% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{read_input}
\alias{read_input}
\title{read_sc_counts
Create seurat object from cellranger counts}
\usage{
read_input(name, input, mincells, minfeats, exp_design)
}
\arguments{
\item{name}{sample name}

\item{input}{path to cellranger counts}

\item{mincells}{min number of cells for which a feature is recorded}

\item{minfeats}{min number of features for which a cell is recorded}

\item{exp_design}{experiment design table}
}
\value{
Seurat object
}
\description{
read_sc_counts
Create seurat object from cellranger counts
}
