% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{recall}
\alias{recall}
\title{Calculates Recall from a measures dataframe}
\usage{
recall(df)
}
\arguments{
\item{df}{dataframe with measures}
}
\value{
calculated metric
}
\description{
Calculates Recall from a measures dataframe
}
\examples{
df <- data.frame(TP = 4, FP = 3, TN = 2, FN = 1)
df$Recall <- recall(df)
}
\keyword{metrics}
