% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{rename_clusters}
\alias{rename_clusters}
\title{rename_clusters
renames seurat clusters according to dictionary, and stores result in new
cell_type metadata column.}
\usage{
rename_clusters(seu, new_clusters = NULL)
}
\arguments{
\item{seu}{Non-annotated seu with markers}

\item{new_clusters}{Vector of names to assign to clusters.}
}
\value{
Annotated seu object
}
\description{
rename_clusters
renames seurat clusters according to dictionary, and stores result in new
cell_type metadata column.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- 1
rename_clusters(seu = pbmc_tiny, new_clusters = "typeA")
}
