% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{rename_samples}
\alias{rename_samples}
\title{Give samples a more descriptive name}
\usage{
rename_samples(exp_design, condition_columns, counts_table)
}
\arguments{
\item{exp_design}{A data frame describing the experimental design}

\item{condition_columns}{An integer vector. Indices of columns to use as
conditions}

\item{counts_table}{Counts table where renaming will be propagated}
}
\value{
A list. Exp design is the renamed experimental design, counts_table
is the renamed counts table
}
\description{
\code{rename_samples} takes an experimental design table and renames samples
with their experimental conditions, making the design more descriptive at
a first glance and easier to interpret
}
