% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{rename_with_dict}
\alias{rename_with_dict}
\title{Use a dictionary to rename a vector
\code{rename_with_dict} renames a vector by consulting a list that links
the original names with the desired names}
\usage{
rename_with_dict(vector, dict)
}
\arguments{
\item{vector}{Vector to rename}

\item{dict}{Dictionary containing name-new_name pairs}
}
\value{
Renamed vector
}
\description{
Use a dictionary to rename a vector
\code{rename_with_dict} renames a vector by consulting a list that links
the original names with the desired names
}
