% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{scale_data_matrix}
\alias{scale_data_matrix}
\title{Scale a matrix using minimum-maximum method}
\usage{
scale_data_matrix(data_matrix, norm_by_col = FALSE)
}
\arguments{
\item{data_matrix}{to be scaled}

\item{norm_by_col}{boolean flag: if true scaling will be performed
by columns intead of by rows. Default: FALSE}
}
\value{
scaled matrix
}
\description{
Scale a matrix using minimum-maximum method
}
\keyword{method}
