% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{scale_range}
\alias{scale_range}
\title{Funtion used to scale a vector of numeric values from
(x_min,x_max) range to
(y_min,y_max) range.
f : R -> R
(x_min, x_max) -> (y_min, y_max)
vect - x_min
f(vect,y_min,y_max) =  ------------- x (y_max - y_min) + y_min
x_max - x_min}
\usage{
scale_range(vect, nmin, nmax)
}
\arguments{
\item{vect}{vector to be transformated}

\item{nmin}{new minimum}

\item{nmax}{new maximum}
}
\value{
transformated vector to new range
}
\description{
Funtion used to scale a vector of numeric values from
(x_min,x_max) range to
(y_min,y_max) range.
f : R -> R
(x_min, x_max) -> (y_min, y_max)
vect - x_min
f(vect,y_min,y_max) =  ------------- x (y_max - y_min) + y_min
x_max - x_min
}
