% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{set_default_width}
\alias{set_default_width}
\title{set_default_width}
\usage{
set_default_width(
  enrich_obj,
  default = 12,
  showCategory = 30,
  threshold = 30,
  character_size = 0.04
)
}
\arguments{
\item{enrich_obj}{Enrichment object.}

\item{default}{Default width, will be modified by new calculations (and NOT
simply replaced). Default 12.}

\item{showCategory}{showCategory. Default 30.}

\item{threshold}{threshold. Default 30.}

\item{character_size}{An integer. Size of characters in plot. Default 0.04.}
}
\value{
Optimal width.
}
\description{
\code{set_default_width} Calculates optimal default with for plot object.
}
\examples{
 \dontrun{
   set_default_width(enrich_obj, default = 12, showCategory = 30,
                     threshold = 30, character_size = 0.04)
 }
}
