% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{sketch_sc_experiment}
\alias{sketch_sc_experiment}
\title{Perform sketching on single-cell experiment}
\usage{
sketch_sc_experiment(
  seu,
  assay = "RNA",
  method = "LeverageScore",
  sketched.assay = "sketch",
  cell.pct = 25,
  force.ncells = NA_integer_,
  features = NULL
)
}
\arguments{
\item{seu}{Seurat object to sketch.
Default value of 5000, recommended by Seurat tutorials.}

\item{assay}{A string. Name of assay to sketch. Default "RNA".}

\item{method}{A string. Method to use for sketching. Default "LeverageScore".}

\item{sketched.assay}{A string. Name of assay where sketch will be saved.}

\item{cell.pct}{A numeric. Percentage of total cells to consider
representative of the experiment. Default 25, as suggested by sketching
tutorial.}

\item{force.ncells}{An integer. Number of cells to forcibly use in sketch.
NA by default.}

\item{features}{A character vector of feature names to include in the
sketched assay.}
}
\value{
A seurat object with a new sketched assay.
}
\description{
\code{sketch_sc_experiment} determines optimal cell number to sketch seurat assay.
If it is larger than specified minimum, it proceeds with sketching, else it
is skipped.
}
\examples{
 \dontrun{
   sketched_experiment <- sketch_sc_experiment(seu = nonsketched_experiment,
          assay = "RNA", method = "LeverageScore", sketched.assay = "sketch",
          cell.pct = 50)
 }
}
