% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{split_str}
\alias{split_str}
\title{split_str}
\usage{
split_str(string, split = NULL)
}
\arguments{
\item{string}{String to split.}

\item{split}{Expression by which to split string.}
}
\value{
Split string.
}
\description{
\code{split_str} is a wrapper for a specific use case of base R strsplit function.
}
\examples{
split_str(string = "one,two", split = ",")
}
