% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{synth_DEGs_by_condition}
\alias{synth_DEGs_by_condition}
\title{Generate DEGs using desired method in specified vectors.}
\usage{
synth_DEGs_by_condition(
  up_vector,
  stat_vector,
  effect_size,
  deg_method = "norm",
  stdev = 0.2
)
}
\arguments{
\item{up_vector}{Samples that will be upregulated}

\item{stat_vector}{Samples that will not be upregulated, but still need
to be randomised in order to not introduce a bias}

\item{effect_size}{Effect size to emulate in expression values}

\item{deg_method}{Method to use. Values: "norm" (\code{synth_normal_DEGs}), det
(\code{synth_deterministic_DEGs}) and "exp" (\code{synth_exponential_DEGs})}

\item{stdev}{Standard deviation for distribution. Used in \code{synth_normal_DEGs}}
}
\value{
Two vectors of values by which to multiply expression levels in
counts table. Upregulated vector includes effect size, static vector only
includes noise from the normal distribution.
}
\description{
\code{synth_DEGs_by_condition} is a wrapper for the synthetic DEG generation
functions. Provide sample vectors and desired method, and this function will
dispatch the corresponding synthetic generator
}
