% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{synth_normal_DEGs}
\alias{synth_normal_DEGs}
\alias{synth_exponential_DEGs}
\title{Generate synthetic DEGs following a random distribution.}
\usage{
synth_normal_DEGs(up_vector, stat_vector, effect_size, stdev = 0.2)

synth_exponential_DEGs(up_vector, stat_vector, effect_size, stdev = NULL)
}
\arguments{
\item{up_vector}{Samples that will be upregulated}

\item{stat_vector}{Samples that will not be upregulated, but still need
to be randomised in order to not introduce a bias}

\item{effect_size}{Effect size to emulate in expression values}

\item{stdev}{Standard deviation for distribution. Used in \code{synth_normal_DEGs}}
}
\value{
Two vectors of values by which to multiply expression levels in
counts table. Upregulated vector includes effect size, static vector only
includes noise from the normal distribution.
}
\description{
\code{synth_normal_DEGs} generates two random vectors by which to multiply
expression values of samples in a counts table. Randomisation is taken from
the normal distribution, with a default standard deviation of two and an
effect size supplied by the user

\code{synth_exponential_DEGs} follows an exponential distribution instead

\code{synth_deterministic_DEGs} is a simple multiplication, with no noise
}
