% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{tag_doublets}
\alias{tag_doublets}
\title{tags barcodes in a seurat object that appear in a vector of doublets}
\usage{
tag_doublets(seu, doublet_list)
}
\arguments{
\item{seu}{Seurat object to tag}

\item{doublet_list}{Vector of barcodes to be marked as doublet. Default NULL}
}
\value{
Seurat object with tagged doublets
}
\description{
tags barcodes in a seurat object that appear in a vector of doublets
}
\examples{
data(pbmc_tiny)
tag_doublets(seu = pbmc_tiny,
             doublet_list = c("ATGCCAGAACGACT", "CATGGCCTGTGCAT"))
}
