% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{tag_qc}
\alias{tag_qc}
\title{tag_qc}
\usage{
tag_qc(seu, minqcfeats = 500, percentmt = 5, doublet_list = NULL)
}
\arguments{
\item{seu}{Seurat object to tag}

\item{minqcfeats}{An integer. Minimum features to consider a cell valid.
Default 500}

\item{percentmt}{A float. Maximum MT percentage to consider a cell valid.
Default 5}

\item{doublet_list}{Vector of barcodes to be marked as doublet. Default NULL.}
}
\value{
Seurat object with tagged metadata
}
\description{
\code{tag_qc} tags cell IDs not passing QC filters
}
\examples{
\dontrun{
  # Integrative experiment, previously loaded doublet list:
  tag_qc(seu = pbmc_tiny, minqcfeats = 500, percentmt = 5,
         doublet_list = c("ATGCCAGAACGACT", "CATGGCCTGTGCAT"))
  # Per-sample experiment, no prior doublet knowledge
  tag_qc(seu = pbmc_tiny, minqcfeats = 500, percentmt = 5,
         doublet_list = NULL) 
}
}
\keyword{preprocessing,}
\keyword{qc}
