% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target}
\alias{target}
\title{Target data containing sample information for the toc data}
\format{
A data.frame of sample information. Rows are samples, columns are different experimental variables/factors
}
\usage{
data(target)
}
\description{
Target data containing sample information for the toc data
}
\examples{
data(target)
}
\keyword{datasets}
