% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toc}
\alias{toc}
\title{Table of counts dummy dataset for DEgeneHunter examples}
\format{
A truncated data.frame of aligned read counts for genes (rows) and
samples (columns)
}
\usage{
data(toc)
}
\description{
Table of counts dummy dataset for DEgeneHunter examples
}
\examples{
data(toc)
}
\keyword{datasets}
