% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{translate_ids_orgdb}
\alias{translate_ids_orgdb}
\title{Translates between gene IDs via orgdb}
\usage{
translate_ids_orgdb(
  ids,
  input_id,
  output_id = "ENTREZID",
  org_db = org_db,
  just_output_ids = FALSE
)
}
\arguments{
\item{ids}{set of IDs to be translated}

\item{input_id}{keytype of input ids}

\item{output_id}{keytype of id to return}

\item{org_db}{org_db to use to obtain translation}

\item{just_output_ids}{return only a vector of the output ids
(if FALSE, returns df including input and output ids)}
}
\value{
data.frame with input and output ids, or just vector of
output ids if just_output_ids set to TRUE
}
\description{
Translates between gene IDs via orgdb
}
\examples{
translate_ids_orgdb(ids = c("ENSG00000197283", "ENSG00000141510"),
                    input_id = "ENSEMBL",
                    output_id = "ENTREZID",
                    org_db = org.Hs.eg.db::org.Hs.eg.db
)
}
