% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_RNA_functions.R
\name{translate_miRNA_ids}
\alias{translate_miRNA_ids}
\title{translate_miRNA_ids}
\usage{
translate_miRNA_ids(
  miRNA_IDs,
  input_keytype = "VW-MIMAT-22.0",
  output_keytype = "NAME"
)
}
\arguments{
\item{miRNA_IDs}{Accession number IDs}

\item{input_keytype}{Input keytype}

\item{output_keytype}{Output keytype}
}
\value{
the miRNA names translate from the accession number IDs
}
\description{
\code{translate_miRNA_ids} get miRNA names from accession numbers
}
\examples{
miRNA_IDs <- c("MIMAT0000158")
translate_miRNA_ids(miRNA_IDs)
}
