% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library_new.R
\name{trycatch_pairwise_termsim}
\alias{trycatch_pairwise_termsim}
\title{if pairwise_termsim throws an error, remove dup cat descriptions
Only seems to be a problem with Reactome - to study further
Further investigation re: number of cats also required}
\usage{
trycatch_pairwise_termsim(enr, num_cats = 200)
}
\arguments{
\item{enr}{enrichment object to be studied}

\item{num_cats}{number of categories to be shown}
}
\value{
enrichment object after add termsim info
}
\description{
if pairwise_termsim throws an error, remove dup cat descriptions
Only seems to be a problem with Reactome - to study further
Further investigation re: number of cats also required
}
