% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{weighted_cont_table}
\alias{weighted_cont_table}
\title{Takes boolean vectors of same length and numeric vector of weights}
\usage{
weighted_cont_table(
  experiment,
  gold_standard,
  weights,
  background_weights = NULL,
  false_weights = 1
)
}
\description{
Takes boolean vectors of same length and numeric vector of weights
}
