% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_abgenes_Hunter.R
\name{write_abgenes_report}
\alias{write_abgenes_report}
\title{write_abgenes_report}
\usage{
write_abgenes_report(
  final_results,
  output_dir = getwd(),
  template_folder = NULL,
  source_folder = NULL,
  p_adj_cutoff = 0.05,
  z_score_cutoff = 3,
  top_N = 10
)
}
\arguments{
\item{final_results}{main_abgenes_Hunter results object.}

\item{output_dir}{Directory where report will be written. Default: working
directory.}

\item{template_folder, source_folder}{paths to htmlreportR templates and
source folders, respectively. Default: NULL.}

\item{p_adj_cutoff}{An integer. Adjusted p-value cutoff to consider a gene
aberrantly expressed. Default 0.05.}

\item{z_score_cutoff}{An integer. Z-score value threshold to tag genes and
samples as aberrant. Default 3.'}

\item{top_N}{Top N values to represent in summary tables. Default: 10.}
}
\value{
None.
}
\description{
\code{write_abgenes_report} writes main_abgenes_Hunter report in html format.
}
\examples{
 \dontrun{
		main_abgenes_Hunter(sample_annotation = sample_annotation,
		 anno_database = anno_database, count_ranges = count_ranges,
      gene_mapping_file = gene_mapping_file, count_files = count_files,
     dataset = dataset, cpu = cpu, fpkm_cutoff = fpkm_cutoff,
	 implementation = implementation, p_adj_cutoff = p_adj_cutoff,
	 max_dim_proportion = max_tested_dimension_proportion, top_N = top_N,
      z_score_cutoff = z_score_cutoff, hpo_file = hpo_file,
	 stats_path = stats_path)
 }
}
