% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{write_annot_output}
\alias{write_annot_output}
\title{write_annot_output
writes final counts matrix and annotated metadata of a single-cell
experiment.}
\usage{
write_annot_output(final_results, opt = NULL, assay = "RNA", layer = "data")
}
\arguments{
\item{final_results}{final_results object output from main_annotate_sc}

\item{opt}{Options used in script call.}

\item{assay, layer}{Assay and layer from where counts will be retrieved.}
}
\value{
invisible(NULL)
}
\description{
write_annot_output
writes final counts matrix and annotated metadata of a single-cell
experiment.
}
\examples{
\dontrun{
   write_annot_output(final_results = test_results, opt = params,
                      assay = "RNA", layer = "data")
}
}
