% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_clusters_to_enrichment}
\alias{write_clusters_to_enrichment}
\title{Write Main clusters to enrichment output
This function allows you to report the Functional analysis.}
\usage{
write_clusters_to_enrichment(
  output_path = "results",
  output_file = "results",
  mode = "PR",
  enrichments_ORA = NULL,
  task_size = 1,
  workers = 1,
  top_categories = NULL,
  group_results = FALSE,
  node_label = "category",
  size_item = 1,
  size_category = 1,
  size_edge = 1,
  hilight = "none",
  hilight_alpha = 0.3,
  n_category = 30,
  sim_thr = 0.7,
  pvalcutoff = 0.1,
  gene_attributes = NULL,
  max_genes = 200,
  summary_common_name = "ancestor",
  simplify = FALSE,
  gene_attribute_name = NULL,
  clean_parentals = FALSE,
  source_folder = NULL,
  template_folder = file.path(find.package("ExpHunterSuite"), "templates"),
  files_css = file.path(template_folder, "styles.css")
)
}
\arguments{
\item{output_path}{output folder}

\item{output_file}{output file name for heatmaps}

\item{mode}{type of output to produce - P for plots, R for reports, and S for summarized heatmaps}

\item{enrichments_ORA}{list of enrich results for all clusters}

\item{task_size}{number of elements per packages used}

\item{workers}{(OPTIONAL) cores for parallel features}

\item{top_categories}{numbers of categories from each cluster to use for merge}

\item{group_results}{experimental - whether to group results in the emap plot}

\item{node_label}{How to label the nodes. Default "category". Other possible
values: "all", "none", "item", "exclusive", "share".}

\item{size_item}{Size of nodes. Smaller than 1 values will reduce size,
higher than 1 values will increase it. Default: 1.}

\item{size_category}{Size of category labels. Smaller than 1 values will
reduce size, higher than 1 values will increase it. Default: 1.}

\item{size_edge}{Size of edges. Smaller than 1 values will reduce size,+
higher than 1 values will increase it. Default: 1.}

\item{hilight}{Name of category to highlight. Default "none".}

\item{hilight_alpha}{Transparency value that will be applied to categories
not highlighted. Must be a value between 0 (completely transparent) and 1
(completely opaque). Default: 0.3.}

\item{n_category}{number of categories in the figures (per cluster)}

\item{sim_thr}{value to use when combining similar categories in summary mode}

\item{pvalcutoff}{used to select terms for summarizing}

\item{gene_attributes}{named list of attributes e.g. FCs for emap plot coloured nodes (genes)}

\item{max_genes}{maximum number of genes to plot in cnet plot}

\item{summary_common_name}{'significant' to use the most significant term to label each summarized group
'ancestor' to use the common ancestor of the group"}

\item{simplify}{Activate for process ClusterProfiler results with simplify function}

\item{gene_attribute_name}{name for the legend in the emap plot for the nodes (genes)}

\item{clean_parentals}{Activate to reduce significant terms in GO by removin the parentals}

\item{source_folder}{htmlreportR source folder to load js and css libraries}

\item{template_folder}{(OPTIONAL) RMD templates folder}

\item{files_css}{Path to css file to inject. By default it injects styles.css
found in ExpHunterSuite's templates directory.}
}
\value{
void
}
\description{
Write Main clusters to enrichment output
This function allows you to report the Functional analysis.
}
\examples{
\dontrun{
   write_clusters_to_enrichment(enrichments_ORA = ORA_results)
}
}
