% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{write_df_list_as_tables}
\alias{write_df_list_as_tables}
\title{Function to write expression package results.
Write to disk the results of each diff expression package}
\usage{
write_df_list_as_tables(df_list, prefix, root = getwd())
}
\arguments{
\item{df_list}{dataframes list}

\item{prefix}{prefix concatenated to output file name}

\item{root}{path where write}
}
\value{
void
}
\description{
Function to write expression package results.
Write to disk the results of each diff expression package
}
\examples{
# Typically used with the @all_counts_for_plotting slot of the main output
deg_res <- list(aa=data.frame(a=1:5,b=1:5), bb=data.frame(a=6:10,b=6:10))
write_df_list_as_tables(deg_res, prefix="test", root=tempdir())
}
\keyword{output}
