% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{write_enrich_files}
\alias{write_enrich_files}
\title{Write enrichment files related to functional_hunter results list}
\usage{
write_enrich_files(func_results, output_path = getwd())
}
\arguments{
\item{func_results}{functional enrichment results}

\item{output_files}{output folder path}
}
\value{
void
}
\description{
Write enrichment files related to functional_hunter results list
}
\examples{
# Use a @functional_hunter or @multienricher result object to 
# create real files
write_enrich_files(list(),"./")
}
\keyword{export}
