% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_expression_data}
\alias{write_expression_data}
\title{Function to create the expression hunter folder
Used by the degenes_Hunter script}
\usage{
write_expression_data(final_results, output_files)
}
\arguments{
\item{final_results}{object containing expression results,
as created by main_degenes_Hunter}

\item{output_files}{Output path}
}
\value{
used for side effect of creating the folder and files
}
\description{
Function to create the expression hunter folder
Used by the degenes_Hunter script
}
\examples{
# Can only be run after running main_degenes_Hunter 
# eg as part of the degenes_Hunter.R script
\dontrun{
     write_expression_data(final_results, opt$output_files)
}
}
