% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_expression_report}
\alias{write_expression_report}
\title{Write Main DEgenes Hunter report}
\usage{
write_expression_report(
  exp_results,
  output_files = getwd(),
  template_folder = NULL,
  opt = NULL,
  source_folder = NULL,
  files_css = file.path(template_folder, "styles.css")
)
}
\arguments{
\item{exp_results}{DEG analysis results object}

\item{output_files}{output file path}

\item{template_folder}{htmlreportR templates folder}

\item{opt}{option list}

\item{source_folder}{Folder where js and css templates to load in htmlreportR
object are located.}

\item{files_css}{Path to css file to inject. By default it injects styles.css
found in ExpHunterSuite's templates directory.}
}
\value{
void
}
\description{
This function allows you to report the DEG analysis.
}
\examples{
\dontrun{
     # Load DE analysis results
     degh_output <- list() # data(degh_output)
     write_expression_report(degh_output)
}
}
