% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{write_opt}
\alias{write_opt}
\title{Write parsed options}
\usage{
write_opt(opt, output = getwd())
}
\arguments{
\item{opt}{Option list to write.}

\item{output}{Path where opt file will be saved.}
}
\value{
invisible(NULL)
}
\description{
\code{write_opt} takes a parsed option list and writes it to disk, in order
to have a control of the latest execution of a script that is easily
readable.
}
\examples{
 \dontrun{
   load_SingleR_Ref(path = "my_reference",
                    filter = "Age <= 3 years | phenotype == patient")
 }
}
