% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{write_sc_report}
\alias{write_sc_report}
\title{write_sc_report
Write integration HTML report}
\usage{
write_sc_report(
  final_results,
  output = getwd(),
  out_suffix = NULL,
  template_folder,
  source_folder = NULL,
  opt,
  params,
  template = NULL,
  use_canvas = TRUE
)
}
\arguments{
\item{final_results}{Output from main_sc_Hunter function}

\item{output}{directory where report will be saved}

\item{out_suffix}{suffix to add to output file name. Useful when rendering
different templates with this function (as is the case in our workflow)}

\item{template_folder}{directory where template is located}

\item{source_folder}{htmlreportR source folder}

\item{opt}{Processed ptions list, will be consulted in report.}

\item{params}{Input options list, will be included as execution parameters.}

\item{template}{Template to render}

\item{use_canvas}{Parameter to select whether or not CanvasXpress plots will
be triggered in templates where this control parameter has been implemented.
Setting it to FALSE can be useful for big datasets, as CanvasXpress might
have trouble in certain plots}
}
\value{
invisible(NULL)
}
\description{
write_sc_report
Write integration HTML report
}
\keyword{preprocessing,}
\keyword{report}
\keyword{write,}
