% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_Hunter.R
\name{write_seurat_report}
\alias{write_seurat_report}
\title{write_seurat_report
Write integration HTML report}
\usage{
write_seurat_report(
  final_results,
  output = getwd(),
  name = NULL,
  template_folder,
  source_folder = NULL,
  target_genes = NULL,
  int_columns = NULL,
  DEG_list = NULL,
  cell_annotation = NULL,
  template = NULL,
  out_name = NULL,
  use_canvas = TRUE
)
}
\arguments{
\item{output}{directory where report will be saved}

\item{name}{experiment name, will be used to build output file name}

\item{template_folder}{directory where template is located}

\item{source_folder}{htmlreportR source folder}

\item{int_columns}{factors present in experiment design}

\item{use_canvas}{Parameter to select whether or not CanvasXpress plots will be
triggered in templates where this control parameter has been implemented.
Setting it to FALSE can be useful for big datasets, as CanvasXpress might
have trouble in certain plots.}

\item{int}{integrated seurat object}
}
\value{
nothing
}
\description{
write_seurat_report
Write integration HTML report
}
\keyword{preprocessing,}
\keyword{report}
\keyword{write,}
