% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{add_gnomAD_AF}
\alias{add_gnomAD_AF}
\alias{add_gnomAD_AF,data.table-method}
\alias{add_gnomAD_AF,GRanges-method}
\title{Add AF to GRanges}
\usage{
add_gnomAD_AF(
  object,
  genome_assembly = "hg19",
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)

\S4method{add_gnomAD_AF}{data.table}(
  object,
  genome_assembly = "hg19",
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)

\S4method{add_gnomAD_AF}{GRanges}(
  object,
  genome_assembly = "hg19",
  max_af_cutoff = 0.001,
  populations = c("AF", "AF_afr", "AF_amr", "AF_eas", "AF_nfe", "AF_popmax"),
  ...
)
}
\arguments{
\item{object}{either a data.table of allelic counts or a GRanges object}

\item{genome_assembly}{one of "hg19", "hs37d5", "hg38", "GRCh38"
It can also be any full string of a MafDb provided by
\code{\link[GenomicScores]{availableGScores}}.}

\item{max_af_cutoff}{cutoff for a variant to be considered rare. Default
0.001}

\item{populations}{The populations to be annotatated.}

\item{...}{Used for backwards compatibility (gene_assembly ->
genome_assembly)}
}
\value{
a data.frame containing original data as well as the minor allele
frequencies
}
\description{
appending the minor allele frequency to GRanges using gnomAD
}
\examples{
file <- system.file("extData", "allelic_counts_HG00187.csv",
package = "ExpHunterSuite", mustWork = TRUE)
maeCounts <- data.table::fread(file)                                                                       
maeRes <- DESeq4MAE(maeCounts)                                                                 
\dontrun{
res <- add_gnomAD_AF(maeRes, genome_assembly = genome_assembly,
populations = "AF")            
}
}
\author{
\href{https://github.com/gagneurlab/tMAE}{gagneurlab}
}
