% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{AE_Sample_Overview}
\alias{AE_Sample_Overview}
\title{Simple wrapper for OUTRIDER::plotVolcano function. Calls it for specified
sample in input Outrider DataSet (ods)}
\usage{
AE_Sample_Overview(ods, sample)
}
\arguments{
\item{ods}{Outrider DataSet object.}

\item{sample}{Sample to plot.}
}
\value{
Volcano plot of aberrantly expressed genes in specified sample
in provided ods.
}
\description{
\code{AE_Sample_Overview} takes a sample and an Outrider DataSet (ods), and
prints a volcano plot of aberrantly expressed genes in said sample.
}
