% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{analyze_sc_query}
\alias{analyze_sc_query}
\title{analyze_sc_query}
\usage{
analyze_sc_query(seu, query, sigfig = 2, sample_col = "sample", layer = "data")
}
\arguments{
\item{seu}{Seurat object.}

\item{query}{Vector of query genes whose expression to analyze.}

\item{sigfig}{Significant figure cutoff.}

\item{sample_col}{Name of column that specifies sample. Default "sample",
as per our workflow.}

\item{layer}{Seurat object layer to analyze.}
}
\value{
A list with the three query analysis objects.
}
\description{
\code{analyze_sc_query} is a wrapper for the three query analysis steps:
\code{get_query_distribution}, \code{get_query_pct} by samples and \code{get_query_pct} by
samples and cell types.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- c(rep(1, 7), rep(2, 8))
analyze_sc_query(pbmc_tiny, c("PPBP", "CA2"), 2, sample_col = "orig.ident")
}
