% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{annotate_clusters
\code{annotate_clusters} renames seurat clusters according to dictionary}
\usage{
annotate_clusters(seu, new_clusters = NULL)
}
\arguments{
\item{seu}{Non-annotated seu with markers}

\item{new_clusters}{Vector of names to assign to clusters.}
}
\value{
Annotated seu object
}
\description{
annotate_clusters
\code{annotate_clusters} renames seurat clusters according to dictionary
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- 1
annotate_clusters(seu = pbmc_tiny, new_clusters = "typeA")
}
