% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{annotate_seurat}
\alias{annotate_seurat}
\title{annotate_seurat}
\usage{
annotate_seurat(
  seu,
  cell_annotation = NULL,
  logfc.threshold = 0.1,
  subset_by = NULL,
  cluster_annotation = NULL,
  p_adj_cutoff = 1e-05,
  verbose = FALSE,
  assay = "RNA",
  integrate = FALSE,
  min.pct = 0.1,
  layer = "scale.data"
)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{cell_annotation}{Table of cell types and their associated markers}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1
Increasing logfc.threshold speeds up the function, but can miss weaker signals.
If the \code{slot} parameter is "scale.data" no filtering is performed.}

\item{subset_by}{Metadata column by which seurat object will be subset for
marker calculation. Default NULL.}

\item{cluster_annotation}{A data frame detailing new names for identified
clusters. Requires prior clustering knowledge.}

\item{p_adj_cutoff}{Minimum adjusted p-value of markers to consider.}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}

\item{assay}{A string. Assay whose markers will be calculated. Default "RNA",
as per usual workflow.}

\item{integrate}{Whether or not integrative analysis is active. Does not
bother with subsetting if FALSE.}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{layer}{Name of layer to get or set}
}
\value{
A seurat object with a new sketched assay.
}
\description{
\code{annotate_seurat} is a wrapper and dispatch for different annotation
strategies for a seurat object.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- c(1, 2)
clust_anno <- data.frame(seurat_clusters = seq(2), name = c("T Cells",
                         "B Cells"))
anno_tiny <- annotate_seurat(seu = pbmc_tiny,
                             cluster_annotation = clust_anno)
}
