% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{annotate_seurat}
\alias{annotate_seurat}
\title{annotate_seurat
is a wrapper and dispatch for different annotation strategies for a seurat
object.}
\usage{
annotate_seurat(
  seu,
  cell_annotation = NULL,
  subset_by = NULL,
  cluster_annotation = NULL,
  p_adj_cutoff = 1e-05,
  verbose = FALSE,
  assay = "RNA",
  integrate = FALSE
)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{cell_annotation}{Table of cell types and their associated markers}

\item{subset_by}{Metadata column by which seurat object will be subset for
marker calculation. Default NULL.}

\item{cluster_annotation}{A data frame detailing new names for identified
clusters. Requires prior clustering knowledge.}

\item{p_adj_cutoff}{Minimum adjusted p-value of markers to consider.}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}

\item{assay}{A string. Assay whose markers will be calculated. Default "RNA",
as per usual workflow.}

\item{integrate}{Whether or not integrative analysis is active. Does not
bother with subsetting if FALSE.}
}
\value{
A seurat object with a new sketched assay.
}
\description{
annotate_seurat
is a wrapper and dispatch for different annotation strategies for a seurat
object.
}
\examples{
 \dontrun{
   sketched_experiment <- sketch_sc_experiment(seu = nonsketched_experiment,
          assay = "RNA", method = "LeverageScore", sketched.assay = "sketch",
          cell.pct = 50)
 }
}
