% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{apply_SingleR}
\alias{apply_SingleR}
\title{apply_SingleR
\code{apply_SingleR} runs the training part of SingleR::SingleR wrapper for both
training and classifying.}
\usage{
apply_SingleR(
  test,
  ref,
  labels,
  method = NULL,
  clusters = NULL,
  genes = "de",
  sd.thresh = 1,
  de.method = "classic",
  de.n = NULL,
  de.args = list(),
  aggr.ref = FALSE,
  aggr.args = list(),
  recompute = TRUE,
  restrict = NULL,
  quantile = 0.8,
  fine.tune = TRUE,
  tune.thresh = 0.05,
  prune = TRUE,
  assay.type.test = "logcounts",
  assay.type.ref = "logcounts",
  check.missing = TRUE,
  num.threads = BiocParallel::bpnworkers(BPPARAM),
  out_dir = getwd(),
  BNPARAM = NULL,
  BPPARAM = BiocParallel::SerialParam(),
  save_trained_object = FALSE,
  load_trained_object = FALSE
)
}
\arguments{
\item{test}{A numeric matrix of single-cell expression values where rows are genes and columns are cells.
Alternatively, a \link[SummarizedExperiment]{SummarizedExperiment} object containing such a matrix.}

\item{ref}{A numeric matrix of (usually normalized and log-transformed) expression values from a reference dataset,
or a \link[SummarizedExperiment]{SummarizedExperiment} object containing such a matrix;
see \code{\link[SingleR]{trainSingleR}} for details.

Alternatively, a list or \link[S4Vectors]{List} of SummarizedExperiment objects or numeric matrices containing multiple references.
Row names may be different across entries but only the intersection will be used, see Details.}

\item{labels}{A character vector or factor of known labels for all samples in \code{ref}.

Alternatively, if \code{ref} is a list, \code{labels} should be a list of the same length.
Each element should contain a character vector or factor specifying the labels for the columns of the corresponding element of \code{ref}.}

\item{method}{Deprecated.}

\item{clusters}{A character vector or factor of cluster identities for each cell in \code{test}.
If set, annotation is performed on the aggregated cluster profiles, otherwise it defaults to per-cell annotation.}

\item{genes, sd.thresh, de.method, de.n, de.args}{Arguments controlling the choice of marker genes used for annotation, see \code{\link[SingleR]{trainSingleR}}.}

\item{aggr.ref, aggr.args}{Arguments controlling the aggregation of the references prior to annotation, see \code{\link[SingleR]{trainSingleR}}.}

\item{recompute}{Deprecated and ignored.}

\item{restrict}{A character vector of gene names to use for marker selection.
By default, all genes in \code{ref} are used.}

\item{quantile}{A numeric scalar specifying the quantile of the correlation distribution to use to compute the score for each label.}

\item{fine.tune}{A logical scalar indicating whether fine-tuning should be performed.}

\item{tune.thresh}{A numeric scalar specifying the maximum difference from the maximum correlation to use in fine-tuning.}

\item{prune}{A logical scalar indicating whether label pruning should be performed.}

\item{assay.type.test}{An integer scalar or string specifying the assay of \code{test} containing the relevant expression matrix,
if \code{test} is a \link[SummarizedExperiment]{SummarizedExperiment} object.}

\item{assay.type.ref}{An integer scalar or string specifying the assay of \code{ref} containing the relevant expression matrix,
if \code{ref} is a \link[SummarizedExperiment]{SummarizedExperiment} object (or is a list that contains one or more such objects).}

\item{check.missing}{Deprecated, use \code{check.missing.test} and \code{check.missing.ref} instead.}

\item{num.threads}{Integer scalar specifying the number of threads to use for index building and classification.}

\item{out_dir}{Directory where trained object will be saved/loaded from.}

\item{BNPARAM}{Deprecated and ignored.}

\item{BPPARAM}{A \link[BiocParallel]{BiocParallelParam} object specifying how parallelization should be performed in other steps,
see \code{?\link[SingleR]{trainSingleR}} and \code{?\link[SingleR]{classifySingleR}} for more details.}

\item{save_trained_object}{A boolean. If TRUE, trained object will be saved
to disk, and further calls to the pipeline will recognize and load it instead
of retraining.
Default FALSE.}

\item{load_trained_object}{A boolean. If TRUE, trained object will be loaded
from disk.}
}
\description{
apply_SingleR
\code{apply_SingleR} runs the training part of SingleR::SingleR wrapper for both
training and classifying.
}
\examples{
 \dontrun{
   data(pbmc_tiny)
   apply_SingleR(pbmc_tiny, tiny_ref)
 }
}
