% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{breakdown_query}
\alias{breakdown_query}
\title{breakdown_query}
\usage{
breakdown_query(
  input,
  query,
  assay = "RNA",
  layer = "scale.data",
  min_counts = 10
)
}
\arguments{
\item{input}{Input object to breakdown. Can be a seurat object or a
single-cell counts matrix.}

\item{query}{Vector of query genes whose expression to analyze.}

\item{assay}{Seurat object assay to break down.}

\item{layer}{Layer of seurat object assay to break down.}

\item{min_counts}{Counts required to consider a gene as expressed.}
}
\value{
A data frame of the proportion of cells (between 0 and 1) that
express each query gene.
}
\description{
\code{breakdown_query} breaks down the expression of a list of query genes by
the specified parameter.
}
\examples{
data(pbmc_tiny)
breakdown_query(input = pbmc_tiny, query = c("PPBP", "CA2"))
}
