% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{breakdown_query}
\alias{breakdown_query}
\title{breakdown_query
\code{breakdown_query} breaks down the expression of a list of query genes by
the specified parameter.}
\usage{
breakdown_query(seu, query, assay = "RNA", layer = "counts")
}
\arguments{
\item{seu}{Seurat object}

\item{query}{Vector of query genes whose expression to analyse.}

\item{assay}{Seurat assay from which to extract data. Default is "RNA",
the default assay.}

\item{layer}{Layer of Seurat object from which to extract data. Default is
"counts", normalised assay data.}

\item{sigfig}{Significant figure cutoff, default 2}
}
\value{
A data frame of the proportion of cells (between 0 and 1) that
express each query gene.
}
\description{
breakdown_query
\code{breakdown_query} breaks down the expression of a list of query genes by
the specified parameter.
}
