% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{breakdown_query}
\alias{breakdown_query}
\title{breakdown_query
breaks down the expression of a list of query genes by
the specified parameter.}
\usage{
breakdown_query(input, query, assay = "RNA", layer = "data")
}
\arguments{
\item{query}{Vector of query genes whose expression to analyse.}

\item{assay}{Seurat object assay to break down}

\item{layer}{Layer of seurat object assay to break down}
}
\value{
A data frame of the proportion of cells (between 0 and 1) that
express each query gene.
}
\description{
breakdown_query
breaks down the expression of a list of query genes by
the specified parameter.
}
\examples{
data(pbmc_tiny)
breakdown_query(seu = pbmc_tiny, query = c("PPBP", "CA2"))
}
