% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{calculate_markers}
\alias{calculate_markers}
\title{calculate_markers}
\usage{
calculate_markers(
  seu,
  subset_by = NULL,
  verbose = FALSE,
  idents = NULL,
  integrate = FALSE,
  min.pct = 0.01,
  logfc.threshold = 0.1,
  assay = "RNA",
  layer = "data"
)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{subset_by}{Metadata column by which seurat object will be subset for
marker calculation. Default NULL.}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}

\item{idents}{Identity class to which to set seurat object before calculating
markers in case conserved mode cannot be triggered.}

\item{integrate}{Whether or not integrative analysis is active. Does not
bother with subsetting if FALSE.}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1
Increasing logfc.threshold speeds up the function, but can miss weaker signals.
If the \code{slot} parameter is "scale.data" no filtering is performed.}

\item{assay}{A string. Assay whose markers will be calculated. Default "RNA",
as per usual workflow.}

\item{layer}{Name of layer to get or set}
}
\value{
Marker or DEG data frame.
}
\description{
\code{calculate_markers} is a wrapper that dynamically decides the best strategy
to find marker genes
in a seurat dataset
}
\examples{
 \dontrun{
   data(pbmc_tiny)
   calculate_markers(seu = pbmc_tiny, subset_by = "groups",
                     verbose = TRUE, idents = "letter.idents")
 }
}
