% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{calculate_markers}
\alias{calculate_markers}
\title{calculate_markers}
\usage{
calculate_markers(
  seu,
  int_columns,
  verbose = FALSE,
  idents = NULL,
  DEG = FALSE,
  integrate = FALSE
)
}
\arguments{
\item{seu}{Seurat object to analyse.}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}

\item{idents}{Identity class to which to set seurat object before calculating
markers in case conserved mode cannot be triggered.}

\item{DEG}{A boolean.
\itemize{
\item \code{TRUE}: Function will calculate differentally expressed genes.
\item \code{FALSE} (the default): Function will calculate cluster markers.
}}
}
\description{
calculate_markers
}
