% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{calculate_markers}
\alias{calculate_markers}
\title{calculate_markers}
\usage{
calculate_markers(
  seu,
  int_columns,
  verbose = FALSE,
  idents = NULL,
  integrate = FALSE,
  min.pct = 0.25,
  logfc.threshold = 0.25
)
}
\arguments{
\item{seu}{Seurat object to analyse.}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}

\item{idents}{Identity class to which to set seurat object before calculating
markers in case conserved mode cannot be triggered.}

\item{min.pct, logfc.threshold}{See ?Seurat::FindAllMarkers}
}
\value{
Marker or DEG data frame.
}
\description{
calculate_markers
}
\examples{
 \dontrun{
   data(pbmc_tiny)
   calculate_markers(seu = pbmc_tiny, int_columns = "groups",
                     verbose = TRUE, idents = "letter.idents")
 }
}
