% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{calculate_markers}
\alias{calculate_markers}
\title{calculate_markers}
\usage{
calculate_markers(
  seu,
  subset_by,
  verbose = FALSE,
  idents = NULL,
  integrate = FALSE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  assay = "RNA"
)
}
\arguments{
\item{seu}{Seurat object to analyse.}

\item{subset_by}{Metadata column by which seurat object will be subset for
marker calculation.}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}

\item{idents}{Identity class to which to set seurat object before calculating
markers in case conserved mode cannot be triggered.}

\item{integrate}{Whether or not integrative analysis is active. Does not
bother with subsetting if FALSE.}

\item{min.pct, logfc.threshold}{See ?Seurat::FindAllMarkers}

\item{assay}{A string. Assay whose markers will be calculated. Default "RNA",
as per usual workflow.}
}
\value{
Marker or DEG data frame.
}
\description{
is a wrapper that dynamically decides the best strategy to find marker genes
in a seurat dataset
}
\examples{
 \dontrun{
   data(pbmc_tiny)
   calculate_markers(seu = pbmc_tiny, subset_by = "groups",
                     verbose = TRUE, idents = "letter.idents")
 }
}
