% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{check_sc_input}
\alias{check_sc_input}
\title{check_sc_input
checks and pre-processes input of main SC analysis functions. Outdated, needs
overhaul.}
\usage{
check_sc_input(
  metadata = NULL,
  DEG_target = NULL,
  integrate = NULL,
  sketch = NULL,
  SingleR_ref = NULL,
  reduce = NULL,
  fine.tune,
  aggr.ref
)
}
\arguments{
\item{metadata}{Seurat object metadata.}

\item{DEG_target}{A data frame describing DEG analysis to perform.}

\item{integrate}{A boolean, or NULL. Whether or not to perform integration.}

\item{sketch}{A boolean, or NULL. Controls sketching of seurat object.}

\item{SingleR_ref}{SingleR_reference to use. Default NULL.}

\item{reduce}{A boolean, or NULL. Whether or not to reduce input for testing.}

\item{fine.tune}{A boolean.
\itemize{
\item \code{TRUE} (the default): Fine-tune statistical model built for transfer.
\item \code{FALSE}: Do not perform fine-tuning..
}}

\item{aggr.ref}{Logical scalar indicating whether references should be aggregated to pseudo-bulk samples for speed, see \code{\link[SingleR]{aggregateReference}}.}
}
\value{
A markers data frame with a new column for cell type assigned to
cluster.
}
\description{
check_sc_input
checks and pre-processes input of main SC analysis functions. Outdated, needs
overhaul.
}
\examples{
new_opt <- check_sc_input(integrate = TRUE, sketch = TRUE, fine.tune = TRUE,
                  aggr.ref = FALSE, SingleR_ref = "Non_NULL", reduce = FALSE)
new_opt
}
