% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_Hunter.R
\name{check_sc_input}
\alias{check_sc_input}
\title{check_sc_input
check input of main SC analysis function}
\usage{
check_sc_input(metadata, DEG_columns)
}
\description{
check_sc_input
check input of main SC analysis function
}
