% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsynth_library.R
\name{custom_synth}
\alias{custom_synth}
\title{custom_synth}
\usage{
custom_synth(
  counts_table,
  exp_design,
  nDEGs,
  columns,
  fixed_DEG_lists,
  fixed_upregulated_DEGs,
  effect_sizes,
  fraction_upregulated,
  overlap_size,
  deg_method,
  samples_per_cond
)
}
\arguments{
\item{output_dir}{Directory where counts matrix and DEG table will be saved.}
}
\value{
A list. Element \code{all} contains the entire results object. Element
\code{counts_matrix} contains just the counts matrix. Element \code{DEGs} contains
a data frame detailing which genes are differentially expressed in dataset.
}
\description{
\code{custom_synth} Is a wrapper for the \code{generateSyntheticData} function
from package \code{compcodeR}.
}
\examples{
B_625_625 <- main_custom_synth(dataset = "B_625_625", n.vars = 12500, 
                                 samples.per.cond = 5, n.diffexp = 1250, 
                                 repl.id = 1, seqdepth = 1e7, 
                                 fraction.upregulated = 0.5, 
                                 between.group.diffdisp = FALSE, 
                                 filter.threshold.total = 1, 
                                 filter.threshold.mediancpm = 0, 
                                 fraction.non.overdispersed = 0)
B_625_625
}
