% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{.has_exclusive_idents}
\alias{.has_exclusive_idents}
\title{.has_exclusive_idents
\code{.has_exclusive_idents} checks whether any condition-identity pairs in
seurat object has less than three occurrences, which makes certain analyses
impossible. Not exported}
\usage{
.has_exclusive_idents(seu, cond, idents)
}
\arguments{
\item{seu}{Seurat object}

\item{cond}{Condition to check}

\item{idents}{Identity class to which to set seurat object before calculating
markers in case conserved mode cannot be triggered.}
}
\value{
A boolean. \code{TRUE} if it contains exclusive pairs, \code{FALSE} otherwise.
}
\description{
.has_exclusive_idents
\code{.has_exclusive_idents} checks whether any condition-identity pairs in
seurat object has less than three occurrences, which makes certain analyses
impossible. Not exported
}
