% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_degsynth.R
\name{fcfunc}
\alias{fcfunc}
\title{Custom function to generate exponentially degradated foldchange range}
\usage{
fcfunc(means, fcmin = 1.4, fcmax = 3, meanlog = 1, sdlog = 0.8)
}
\arguments{
\item{means}{means vector}

\item{fcmin}{minimum foldchange}

\item{fcmax}{maximum foldchange}

\item{meanlog}{param of rlnorm function}

\item{sdlog}{param of rlnorm function}
}
\value{
a vector of foldchanges to be applied
}
\description{
Custom function to generate exponentially degradated foldchange range
}
