% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{filter_sc_counts}
\alias{filter_sc_counts}
\title{Filter single-cell counts matrix}
\usage{
filter_sc_counts(object, layers = "data", min_counts)
}
\arguments{
\item{object}{An object}

\item{layers}{A string vector of layers to filter.}

\item{min_counts}{Minimum counts threshold to consider a cell expresses
a gene.}
}
\value{
Seurat object with filtered layer.
}
\description{
\code{filter_sc_counts} takes a seurat object, extracts the counts matrix from
the specified layer and applies a min count filter, then slots it back in
the original seurat layer.
}
\examples{
data(pbmc_tiny)
counts <- SeuratObject::GetAssayData(object = seu, layer = "data")
filtered_pbmc <- filter_sc_counts(object = seu, layer = "data",
                                  min_counts = 2)
filtered_counts <- SeuratObject::GetAssayData(object = seu, layer = "data")
print(counts)
print(filtered_counts)
}
