% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{find_doublets}
\alias{find_doublets}
\title{find_doublets
\code{find_doublets} is a wrapper for the recommended steps for doublet
calculation in package DoubletFinder}
\usage{
find_doublets(seu)
}
\arguments{
\item{seu}{Seurat object}
}
\description{
find_doublets
\code{find_doublets} is a wrapper for the recommended steps for doublet
calculation in package DoubletFinder
}
