% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{findEncodingDim}
\alias{findEncodingDim}
\title{find OUTRIDER encoding dimensions}
\usage{
findEncodingDim(
  ods,
  params = seq(2, min(100, ncol(ods) - 1, nrow(ods) - 1), 2),
  freq = 0.01,
  zScore = 3,
  sdlog = log(1.6),
  lnorm = TRUE,
  inj = "both",
  ...,
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{ods}{An OutriderDataSet}

\item{params}{Set of possible q values.}

\item{freq}{Frequency of outlier, defaults to 1E-2}

\item{zScore}{Set of possible injection Z-score, defaults to 3.}

\item{sdlog}{Standard deviation of the sitribution on the log scale.}

\item{lnorm}{If TRUE, the default, Z-scores are drawn from a log normal
distribution with a mean of \code{log(zScore)} in log-scale.}

\item{inj}{Injection strategy, by default 'both'.}

\item{...}{Further arguments passed on to the \code{controlForConfounders}
function.}

\item{BPPARAM}{BPPARAM object by default bpparam().}
}
\value{
The optimal encoding dimension
}
\description{
\code{findEncodingDim} Finds the optimal encoding dimension for a given data set
by running a grid search based on the provided parameter set.
}
\details{
This function has been copied from version 1.20.1 of package
OUTRIDER, as it has been deleted in more modern versions. OUTRIDER 1.26.2
has supposedly implemented a faster algorithm, but as 21/10/2025 it has not
been incorporated into the DROP pipeline. Thus, we have been forced to
copy the function until we have time to incorporate the new algorithm.
Supposedly, OUTRIDER::findInjectZscore is the function we're looking for,
but we cannot test it right now.
}
\examples{
ods <- OUTRIDER::makeExampleOutriderDataSet()
encDimSearchParams <- c(5, 8, 10, 12, 15)
zScoreParams <- c(2, 3, 5, 'lnorm')
implementation <- 'autoencoder'
register(BiocParallel::MulticoreParam(4))
\dontshow{
    ods <- ods[1:12,1:12]
    encDimSearchParams <- c(2)
    zScoreParams <- c('lnorm')
    register(BiocParallel::SerialParam())
    implementation <- 'pca'
}
ods1 <- OUTRIDER::findEncodingDim(ods, params=encDimSearchParams, 
        implementation=implementation)
OUTRIDER::plotEncDimSearch(ods1)

ods2 <- OUTRIDER::findInjectZscore(ods, zScoreParams=zScoreParams,
        encDimParams=encDimSearchParams, implementation=implementation)
OUTRIDER::plotEncDimSearch(ods2)

}
