% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_clusters_count}
\alias{get_clusters_count}
\title{get_clusters_count}
\usage{
get_clusters_count(results_WGCNA)
}
\arguments{
\item{results_WGCNA}{WGCNA results object.}
}
\value{
number of clusters in a WGCNA object
}
\description{
\code{get_clusters_count} count how many clusters are found in a WGCNA object
}
\examples{
 \dontrun{
   get_clusters_count(results_WGCNA)
 }
}
