% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_clusters_distribution}
\alias{get_clusters_distribution}
\title{get_clusters_distribution
\code{get_clusters_distribution} calculates the percentage of cells that make up
each cluster for each different sample in a seurat object. If clusters are
annotated, it will show cell types instead of cluster number.}
\usage{
get_clusters_distribution(seu, sigfig = 3)
}
\arguments{
\item{seu}{Clustered seurat object.}

\item{sigfig}{Significant figure cutoff}
}
\value{
A data frame with cell type distribution in each sample.
}
\description{
get_clusters_distribution
\code{get_clusters_distribution} calculates the percentage of cells that make up
each cluster for each different sample in a seurat object. If clusters are
annotated, it will show cell types instead of cluster number.
}
