% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_DEG_table}
\alias{get_DEG_table}
\title{Extract and calculate relevant info from DEG table}
\usage{
get_DEG_table(
  input_DEGs,
  col_vector = c("gene", "p_val", "p_val_adj", "avg_log2FC", "pct.1", "pct.2",
    "cause_for_rejection", "DEG"),
  DE_method = NULL
)
}
\arguments{
\item{input_DEGs}{Input data frame, created by our sc_Hunter pipeline.}

\item{col_vector}{Vector of columns to retrieve from table. Default: gene,
p_val, p_val_adj, avg_log2FC, pct.1, pct.2, cause_for_rejection, DEG.}

\item{DE_method}{Method used to calculate DEGs. It will be part of column
names if DE dataframe has been ran through tag_DEGs.}
}
\value{
A subset of the DEG table.
}
\description{
\code{get_DEG_table} extracts relevant columns from a DEG output table and
adds a "prevalent" field.
}
\examples{
stop("This example is missing")
}
