% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_expression_metrics}
\alias{get_expression_metrics}
\title{get_expression_metrics}
\usage{
get_expression_metrics(seu, sigfig, sample_col = "sample", min_counts = 10)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{sigfig}{Significant figure cutoff.}

\item{sample_col}{Seurat object metadata column containing sample
information. Default "sample", as per our workflow.}

\item{min_counts}{An integer. Minimum gene counts to consider a gene as
expressed.}
}
\value{
A list containing expression quality metrics and cluster
distribution metrics.'
}
\description{
\code{get_expression_metrics} is a wrapper for the expression metrics extraction
step of our pipeline.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- c(rep(1, 7), rep(2, 8))
get_expression_metrics(seu = pbmc_tiny, sigfig = 2, sample_col = "groups")
}
