% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_expression_metrics}
\alias{get_expression_metrics}
\title{get_expression_metrics}
\usage{
get_expression_metrics(seu, sigfig)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{sigfig}{Significant figure cutoff.}
}
\value{
A list containing expression quality metrics and cluster
distribution metrics.'
}
\description{
\code{get_expression_metrics} is a wrapper for the expression metrics extraction
step of our pipeline.
}
\examples{
 \dontrun{
   get_expression_metrics(seu = seu, sigfig = 2)
 }
}
