% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{get_ods_results}
\alias{get_ods_results}
\title{Wrapper for DROP OUTRIDER results script.
\code{get_ods_results} extracts the results from an ods object.}
\usage{
get_ods_results(
  ods,
  p_adj_cutoff,
  z_score_cutoff,
  gene_mapping_file,
  sample_anno
)
}
\arguments{
\item{ods}{A fitted outrider dataset.}

\item{p_adj_cutoff}{Adjusted p-value cutoff. Gene adjusted p-value must be
lower than this number in order to be considered aberrantly expressed.}

\item{z_score_cutoff}{Cutoff that gene zScore must surpass in order to be
considered aberrantly expressed.}

\item{gene_mapping_file}{Gene mapping file generated by preprocess_gtf
function.}

\item{sample_anno}{Sample annotation data frame.}
}
\value{
A list with two items. The first one, named "all", contains all
results. The second one, "table", only contains significant results.
}
\description{
Wrapper for DROP OUTRIDER results script.
\code{get_ods_results} extracts the results from an ods object.
}
