% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{get_organism_table}
\alias{get_organism_table}
\title{Table with information abaut all organism available}
\usage{
get_organism_table(
  file = file.path(find.package("ExpHunterSuite"), "external_data",
    "organism_table.txt")
)
}
\arguments{
\item{file}{to be loaded. Default: internal organism table}
}
\value{
organism table
}
\description{
Table with information abaut all organism available
}
\examples{
ot <- get_organism_table()
}
\keyword{method}
