% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_qc_pct}
\alias{get_qc_pct}
\title{get_qc_pct}
\usage{
get_qc_pct(
  seu,
  top = 20,
  assay = "RNA",
  layer = "counts",
  sigfig = 2,
  sample_col = "sample",
  min_counts = 10
)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{top}{Top N genes to retrieve.}

\item{assay}{Seurat object assay to break down.}

\item{layer}{Layer of seurat object assay to break down.}

\item{sigfig}{Significant figure cutoff.}

\item{sample_col}{Seurat object metadata column containing sample
information. Default "sample", as per our workflow.}

\item{min_counts}{Counts required to consider a gene as expressed.}
}
\value{
Gene expression matrix of union of top N expressed genes in all
samples.
}
\description{
\code{get_qc_pct} creates a gene expression matrix of the union of the top N genes
expressed in every sample in a seurat object.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- c(rep(1, 7), rep(2, 8))
get_qc_pct(seu = pbmc_tiny, top = 5, assay = "RNA", layer = "counts",
           sample_col = "orig.ident", sigfig = 2)
}
