% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_qc_pct}
\alias{get_qc_pct}
\title{get_qc_pct
\code{get_qc_pct} creates a gene expressio matrix of the union of the top N genes
expressed in every sample in a seurat object.}
\usage{
get_qc_pct(seu, top = 20, assay = "RNA", layer = "counts", by, sigfig = 2)
}
\arguments{
\item{seu}{Seurat object}

\item{top}{Top N genes to take from each sample.}

\item{assay}{Seurat assay from which to extract data. Default is "RNA",
the default assay.}

\item{layer}{Layer of Seurat object from which to extract data. Default is
"counts", normalised assay data.}

\item{sigfig}{Significant figure cutoff, default 2}
}
\description{
get_qc_pct
\code{get_qc_pct} creates a gene expressio matrix of the union of the top N genes
expressed in every sample in a seurat object.
}
