% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_query_distribution}
\alias{get_query_distribution}
\title{get_query_distribution
\code{get_query_distribution} builds a table of query genes expression levels
across all samples of a Seurat object}
\usage{
get_query_distribution(seu, query, sigfig = 3, sample_col = "sample")
}
\arguments{
\item{seu}{Seurat object}

\item{query}{Vector of query genes whose expression to analyse.}

\item{sigfig}{Significant figure cutoff}

\item{sample_col}{Name of column that specifies sample. Default "sample",
as per our workflow.}
}
\value{
A data frame with expression levels for query genes in each sample.
}
\description{
get_query_distribution
\code{get_query_distribution} builds a table of query genes expression levels
across all samples of a Seurat object
}
\examples{
data(pbmc_tiny)
get_query_distribution(pbmc_tiny, c("PPBP", "CA2"), 2, "orig.ident")
}
