% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_query_pct}
\alias{get_query_pct}
\title{get_query_pct
\code{get_query_pct} gets the percentage of cells in each sample of a seurat
object which expresses genes specified in a list of queries.}
\usage{
get_query_pct(seu, query, by, sigfig = 2, assay = "RNA", layer = "counts")
}
\arguments{
\item{seu}{Seurat object}

\item{query}{Vector of query genes whose expression to analyse.}

\item{sigfig}{Significant figure cutoff}

\item{assay}{Seurat assay from which to extract data. Default is "RNA",
the default assay.}

\item{layer}{Layer of Seurat object from which to extract data. Default is
"counts", normalised assay data.}
}
\value{
A data frame with expression levels for query genes in each sample.
}
\description{
get_query_pct
\code{get_query_pct} gets the percentage of cells in each sample of a seurat
object which expresses genes specified in a list of queries.
}
