% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_query_pct}
\alias{get_query_pct}
\title{get_query_pct
\code{get_query_pct} gets the percentage of cells in each sample of a seurat
object which expresses genes specified in a list of queries.}
\usage{
get_query_pct(seu, query, by, sigfig = 2, assay = "RNA", layer = "counts")
}
\arguments{
\item{query}{Vector of query genes whose expression to analyse.}

\item{sigfig}{Significant figure cutoff}
}
\value{
A data frame with expression levels for query genes in each sample.
}
\description{
get_query_pct
\code{get_query_pct} gets the percentage of cells in each sample of a seurat
object which expresses genes specified in a list of queries.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- c(rep(1, 7), rep(2, 8))
get_query_pct(pbmc_tiny, c("PPBP", "CA2"), "seurat_clusters", 2)
}
