% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_query_pct}
\alias{get_query_pct}
\title{get_query_pct
gets the percentage of cells in each sample of a seurat
object which expresses genes specified in a list of queries.}
\usage{
get_query_pct(seu, query, by, sigfig = 2, assay = "RNA", layer = "data")
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{query}{Vector of query genes whose expression to analyze.}

\item{by}{Factor (metadata column name) by which query will be broken down.}

\item{sigfig}{Significant figure cutoff.}

\item{assay}{Seurat object assay to break down.}

\item{layer}{Layer of seurat object assay to break down.}
}
\value{
A data frame with percentage of cells expressing query genes in each
sample.
}
\description{
get_query_pct
gets the percentage of cells in each sample of a seurat
object which expresses genes specified in a list of queries.
}
\examples{
data(pbmc_tiny)
pbmc_tiny$seurat_clusters <- c(rep(1, 7), rep(2, 8))
get_query_pct(pbmc_tiny, c("PPBP", "CA2"), "seurat_clusters", 2)
}
