% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_sc_markers}
\alias{get_sc_markers}
\title{get_sc_markers
\code{get_sc_markers} performs differential expression analysis on OR selects
conserver markers from a seurat object.}
\usage{
get_sc_markers(seu, cond = NULL, subset_by, DEG = FALSE, verbose = FALSE)
}
\arguments{
\item{seu}{Seurat object to analyse.}

\item{cond}{A string. Condition by which to perform DEG analysis, or by which
to group data to find conserved markers.}

\item{subset_by}{Metadata column by which seurat object will be subset for
marker calculation.}

\item{DEG}{A boolean.
\itemize{
\item \code{TRUE}: Function will calculate differentally expressed genes.
\item \code{FALSE} (the default): Function will calculate cluster markers.
}}

\item{verbose}{A boolean. Will be passed to Seurat function calls.}
}
\value{
A list containing one marker or DEG data frame per cluster, plus
an additional one for global DEGs if performing differential analysis.
}
\description{
get_sc_markers
\code{get_sc_markers} performs differential expression analysis on OR selects
conserver markers from a seurat object.
}
