% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{get_top_genes}
\alias{get_top_genes}
\title{get_top_genes
extracts the top N genes expressed in the highest percentage
of cells for each sample in a seurat object, and returns a vector with the
union of these genes.}
\usage{
get_top_genes(
  seu,
  top = 20,
  assay = "RNA",
  layer = "counts",
  sample_col = "sample"
)
}
\arguments{
\item{seu}{Seurat object to analyze}

\item{top}{Top N genes to retrieve.}

\item{assay}{Seurat object assay to break down}

\item{layer}{Layer of seurat object assay to break down}

\item{sample_col}{Seurat object metadata column containing sample
information. Default "sample", as per our workflow.}
}
\value{
A vector containing the union of the top N genes of each sample of
input Seurat object.
}
\description{
get_top_genes
extracts the top N genes expressed in the highest percentage
of cells for each sample in a seurat object, and returns a vector with the
union of these genes.
}
\examples{
data(pbmc_tiny)
get_top_genes(seu = pbmc_tiny, top = 5, assay = "RNA", layer = "counts",
              sample_col = "orig.ident")
}
