% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{load_DEG_output}
\alias{load_DEG_output}
\title{load output from previous execution of DEG module}
\usage{
load_DEG_output(
  targets,
  load_path,
  min_avg_log2FC,
  p_val_cutoff,
  recalc_query = FALSE,
  min_cell_proportion,
  seu,
  query = NULL,
  DE_method = "wilcox"
)
}
\arguments{
\item{targets}{Vector of target names.}

\item{load_path}{Path where target executions will be found.}

\item{min_avg_log2FC}{Average log2fc cutoff for significant DEGs.
Default 0.5.}

\item{p_val_cutoff}{Adjusted p-val cutoff for significant DEGs.
Default 5e-3.}

\item{recalc_query}{A boolean.
\itemize{
\item \code{TRUE}: Recalculate query metrics.
\item \code{FALSE} (the default): Load query metrics.
}}

\item{min_cell_proportion}{Minimum threshold of percentage of cells
expressing DEG. Despite Seurat::FindMarkers being called "min.pct", it is
not a percentage but a value from 0 to 1.}

\item{seu}{A seurat object. Only used if recalc_query is TRUE.}

\item{query}{A string vector. Query to explore in DEG results. Will only
be used if it differs from query used in saved analysis.}

\item{DE_method}{Method that was used when calculating DEGs. Our pipeline
saves results in a different path depending on the method, as to allow
for multiple simultaneous executions and integration. Default: "wilcox"}
}
\value{
A DEG results list with identical structure to main_sc_Hunter
output.
}
\description{
\code{load_DEG_output} loads a previous execution of our Single-Cell DEA.
}
\examples{
\dontrun{
   load_DEG_output(targets = c("42_vs_68", "Ctrl_vs_Mut"),
                   load_path = "path/to/results")
}
}
