% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_handling.R
\name{load_hunter_folder}
\alias{load_hunter_folder}
\title{Loads stored results in DEgenes Hunter expression analysis
folder and returns in correct object format}
\usage{
load_hunter_folder(path = NULL)
}
\arguments{
\item{path}{of DGH folder}
}
\value{
hunter expression analysis object loaded
}
\description{
Loads stored results in DEgenes Hunter expression analysis
folder and returns in correct object format
}
\examples{
# Used to load a DE analysis result folder
# Related to ouputu generated by @main_degenes_Hunter  
hf <- load_hunter_folder() 
# Will return an empty object, you need a results object
}
\keyword{input}
