% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{load_SingleR_ref}
\alias{load_SingleR_ref}
\title{Load SingleR reference}
\usage{
load_SingleR_ref(path, filter = "")
}
\arguments{
\item{path}{Path to reference to load.}

\item{filter}{Filter to apply to reference.}
}
\value{
Loaded and optionally filtered SingleR reference.
}
\description{
\code{load_SingleR_ref} reads a SingleR reference from specified path. It can
handle multiple versions of the same reference, and also filter it by
metadata.
}
\examples{
 \dontrun{
   load_SingleR_Ref(path = "my_reference",
                    filter = "Age <= 3 years | phenotype == patient")
 }
}
