% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{main_analyze_sc_query}
\alias{main_analyze_sc_query}
\title{analyze_sc_query
is a wrapper for the three query analysis steps:
\code{get_query_distribution}, \code{get_query_pct} by samples and \code{get_query_pct} by
samples and cell types.}
\usage{
main_analyze_sc_query(
  seu,
  query,
  sigfig = 2,
  layer = "counts",
  sample_col = "sample",
  min_counts = 10
)
}
\arguments{
\item{seu}{Seurat object.}

\item{query}{Vector of query genes whose expression to analyze.}

\item{sigfig}{Significant figure cutoff.}

\item{layer}{Seurat object layer to analyze.}

\item{sample_col}{Name of column that specifies sample. Default "sample",
as per our workflow.}

\item{min_counts}{Minimum counts to consider a gene is expressed}
}
\value{
A list with the three query analysis objects.
}
\description{
analyze_sc_query
is a wrapper for the three query analysis steps:
\code{get_query_distribution}, \code{get_query_pct} by samples and \code{get_query_pct} by
samples and cell types.
}
\details{
As this function is supposed to be used after our annotation
pipeline, which has already applied a min_counts filter, default for this
function is 1.
}
\examples{
\dontrun{
   data(pbmc_tiny)
   pbmc_tiny$seurat_clusters <- c(rep(1, 7), rep(2, 8))
   main_analyze_sc_query(pbmc_tiny, c("PPBP", "CA2"), 2,
   sample_col = "orig.ident", layer = "counts")
}
}
