% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{main_annotate_sc}
\alias{main_annotate_sc}
\title{main_annotate_sc
is the main seurat analysis function, which can be used for integrative or
non-integrative analysis.}
\usage{
main_annotate_sc(
  seu,
  minqcfeats = 500,
  percentmt = 5,
  sigfig = 2,
  resolution = 0.5,
  p_adj_cutoff = 0.005,
  name = NULL,
  integrate = FALSE,
  cluster_annotation = NULL,
  cell_annotation = NULL,
  scalefactor = 10000,
  hvgs = 2000,
  subset_by = NULL,
  ndims = 10,
  normalmethod = "LogNormalize",
  verbose = FALSE,
  output = getwd(),
  reduce = FALSE,
  min_cells_per_sample = 500,
  ref_label = NULL,
  SingleR_ref = NULL,
  ref_de_method = NULL,
  ref_n = NULL,
  BPPARAM = SerialParam(),
  doublet_list = NULL,
  k_weight = 100,
  integration_method = "Harmony",
  sketch = FALSE,
  sketch_pct = 25,
  force_ncells = NA_integer_,
  sketch_method = "LeverageScore",
  min.pct = 0.1,
  doublet_path = getwd(),
  min_cell_proportion = 0.1,
  logfc.threshold = 0.25,
  aggr.ref = FALSE,
  fine.tune = TRUE,
  min_counts = 0.5
)
}
\arguments{
\item{seu}{A seurat object.}

\item{minqcfeats}{An integer. Minimum features to consider a cell valid.
Default 500}

\item{percentmt}{A float. Maximum MT percentage to consider a cell valid.
Default 5}

\item{sigfig}{An integer. Significant figures to retain in output.}

\item{resolution}{An integer. Controls clustering granularity. Default 0.5}

\item{p_adj_cutoff}{A float. Adjusted p-value cutoff by which to consider a
marker valid for cell type annotation.}

\item{name}{Project name. Default NULL (no project name).}

\item{integrate}{A boolean.
\itemize{
\item \code{TRUE}: Integrate seurat object using the \code{harmony} package.
\item \code{FALSE} (the default): Do not perform integration.
}}

\item{cluster_annotation}{A data frame. Table to use to rename clusters.}

\item{cell_annotation}{A data frame. Table of markers to use for cell type
annotation.}

\item{scalefactor}{An integer. Factor by which to scale data in normalisation}

\item{hvgs}{An integer. Number of highly-variable features to select.
default 2000.}

\item{subset_by}{A string vector. Categories to consider in integrative
analysis.}

\item{ndims}{An integer. Target dimensions in dimensionality reduction.
Default 10.}

\item{normalmethod}{A string. Method to use in normalisation. Default is
"LogNormalize", the Seurat default.}

\item{verbose}{A boolean.
\itemize{
\item \code{TRUE}: Prints progress bars and messages to closely monitor progress.
\item \code{FALSE} (the default): Print minimal progress messages.
}}

\item{output}{A string. Path to output directory.}

\item{reduce}{A boolean.
\itemize{
\item \code{TRUE}: Skip QC filtering. Intended for development and testing.
\item \code{FALSE} (the default): QC filtering will be performed.
}}

\item{min_cells_per_sample}{An integer. Samples with fewer than this value of
cells will be tagged as low_cell.'}

\item{ref_label}{Reference column to use in annotation.}

\item{SingleR_ref}{SummarizedExperiment object to use as reference
for SingleR cell type annotation. If NULL (default value),
SingleR will not be used.}

\item{ref_de_method}{Method to use for marker calculation in single-cell
reference.}

\item{ref_n}{Top N reference markers to consider in annotation. Higher values
provide a more accurate annotation, but increase noise and computational
time. Will not be used if ref_de_method is empty.}

\item{BPPARAM}{Parameters to pass to BiocParallel framework.}

\item{doublet_list}{A vector containing barcodes to be marked as doublets.
NULL by default. Per-sample analysis finds this vector for every sample,
integrative mode requires this vector.}

\item{k_weight}{Number of neighbors to consider when weighting anchors. Used
in integration.}

\item{integration_method}{A string. Method to use in integration. "Harmony"
(the default), "RPCA", "CCA", "FastMNN" or "scVI".}

\item{sketch}{A boolean. If TRUE, data will be sketched. If FALSE
(the default), data will be analysed as-is.}

\item{sketch_pct}{A numeric. Percentage of total cells to consider
representative of the experiment. Default 12, as suggested by sketching
tutorial.}

\item{force_ncells}{an integer. Sketch function will be forced to select
this many cells.}

\item{sketch_method}{A string. Method to use in score calculation for
sketching.}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{doublet_path}{A string. Path where cell IDs identified as doublets
will be written.}

\item{min_cell_proportion}{Minimum threshold of percentage of cells
expressing DEG. Despite Seurat::FindMarkers being called "min.pct", it is
not a percentage but a value from 0 to 1.}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1
Increasing logfc.threshold speeds up the function, but can miss weaker signals.
If the \code{slot} parameter is "scale.data" no filtering is performed.}

\item{aggr.ref}{Logical scalar indicating whether references should be aggregated to pseudo-bulk samples for speed, see \code{\link[SingleR]{aggregateReference}}.}

\item{fine.tune}{A boolean.
\itemize{
\item \code{TRUE} (the default): Fine-tune statistical model built for transfer.
\item \code{FALSE}: Do not perform fine-tuning..
}}

\item{min_counts}{An integer. Minimum gene counts to consider a gene as
expressed.}
}
\value{
final results list. Contains multiple items:
\itemize{
\item qc: seurat object prior to filtering and analysis.
\item seu: processed seurat object.
\item markers: Data frame of marker genes.
\item sample_qc_pct: Gene expression matrix of union of top N expressed genes in
all samples.
\item integrate: Whether or not to trigger integrative analysis. Workflow
contains a few key differences depending on this argument.
\item clusters_pct: A data frame with the cluster composition of samples.
\item SingleR_annotation: Trained SingleR annotation object.
}
}
\description{
main_annotate_sc
is the main seurat analysis function, which can be used for integrative or
non-integrative analysis.
}
\examples{
 \dontrun{
   main_annotate_sc(seu = seurat_object, minqcfeats = 500, percenmt = 5,
                  query = "TREM2", sigfig = 2, resolution = 0.5,
                  p_adj_cutoff = 5e-3, name = "project_name",
                  integrate = TRUE, cluster_annotation = NULL,
                  cell_annotation = cell_types,
                  scalefactor = 10000, hvgs = 2000, subset_by = "genotype",
                  normalmethod = "LogNormalize", ndims = 10, verbose = FALSE,
                  output = getwd(), reduce = FALSE,
                  ref_label = NULL, SingleR_ref = NULL, ref_de_method = NULL,
                  ref_n = NULL, BPPARAM = NULL, doublet_list = NULL,
                  integration_method = "Harmony", sketch_pct = 12,
                  DEG_p_val_cutoff = 5e-3, min_cells_per_sample = 500,
                  min_avg_log2FC = 0.5, k_weight = 90)
 }
}
