% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{main_annotate_sc}
\alias{main_annotate_sc}
\title{main_annotate_sc
is the main seurat analysis function, which can be used for integrative or
non-integrative analysis.}
\usage{
main_annotate_sc(
  seu,
  minqcfeats,
  percentmt,
  query,
  sigfig = 2,
  resolution = 0.5,
  p_adj_cutoff = 0.005,
  name = NULL,
  integrate = FALSE,
  cluster_annotation = NULL,
  cell_annotation = NULL,
  scalefactor = 10000,
  hvgs = 2000,
  subset_by = NULL,
  ndims = 10,
  normalmethod = "LogNormalize",
  verbose = FALSE,
  output = getwd(),
  reduce = FALSE,
  ref_label = NULL,
  SingleR_ref = NULL,
  ref_de_method = NULL,
  ref_n = NULL,
  BPPARAM = NULL,
  doublet_list = NULL,
  integration_method = "Harmony",
  sketch = FALSE,
  sketch_pct = 25,
  k_weight = 100,
  force_ncells = NA_integer_,
  sketch_method = "LeverageScore",
  doublet_path = getwd()
)
}
\arguments{
\item{seu}{A seurat object.}

\item{minqcfeats}{An integer. Minimum features to consider a cell valid}

\item{percentmt}{A float. Maximun MT percentage to consider a cell valid}

\item{query}{A string vector. List of genes to explore in dataset}

\item{sigfig}{An integer. Significant figures to output}

\item{resolution}{An integer. Controls clustering granularity. Default 0.5}

\item{p_adj_cutoff}{A float. Adjusted p-value cutoff by which to consider a
marker valid for cell type annotation.}

\item{name}{Project name. Default NULL (no project name)}

\item{integrate}{A boolean.
\itemize{
\item \code{TRUE}: Integrate seurat object using the \code{harmony} package.
\item \code{FALSE} (the default): Do not perform integration.
}}

\item{cluster_annotation}{A data frame. Table to use to rename clusters.}

\item{cell_annotation}{A data frame. Table of markers to use for cell type
annotation.}

\item{scalefactor}{An integer. Factor by which to scale data in normalisation}

\item{hvgs}{An integer. Number of highly-variable features to select.
default 2000.}

\item{subset_by}{A string vector. Categories to consider in integrative
analysis.}

\item{ndims}{An integer. Target dimensions in dimensionality reduction.
Default 10.}

\item{normalmethod}{A string. Method to use in normalisation. Default is
"LogNormalize", the Seurat default.}

\item{verbose}{A boolean.
\itemize{
\item \code{TRUE}: Prints progress bars and messages to closely monitor progress.
\item \code{FALSE} (the default): Print minimal progress messages.
}}

\item{output}{A string. Path to output directory.}

\item{reduce}{A boolean.
\itemize{
\item \code{TRUE}: Skip QC filtering. Intended for development and testing.
\item \code{FALSE} (the default): QC filtering will be performed.
}}

\item{ref_label}{Reference column to use in annotation.}

\item{SingleR_ref}{SummarizedExperiment object to use as reference
for SingleR cell type annotation. If NULL (default value),
SingleR will not be used.}

\item{ref_de_method}{Method to use for marker calculation in single-cell
reference.}

\item{ref_n}{Top N reference markers to consider in annotation. Higher values
provide a more accurate annotation, but increase noise and computational
time. Will not be used if ref_de_method is empty.}

\item{BPPARAM}{Parameters to pass to BiocParallel framework.}

\item{doublet_list}{A vector containing barcodes to be marked as doublets.
NULL by default. Per-sample analysis finds this vector
for every sample, integrative mode requires this vector.}

\item{integration_method}{A string. Method to use in integration. "Harmony"
(the default), "RPCA", "CCA", "FastMNN" or "scVI".}

\item{sketch}{A boolean. If TRUE, data will be sketched. If FALSE
(the default), data will be analysed as-is.}

\item{sketch_pct}{A numeric. Percentage of total cells to consider
representative of the experiment. Default 12, as suggested by sketching
tutorial.}

\item{k_weight}{Number of neighbors to consider when weighting anchors. Used
in integration.}

\item{force_ncells}{an integer. Sketch function will be forced to select
this many cells.}

\item{sketch_method}{A string. Method to use in score calculation for
sketching.}

\item{doublet_path}{A string. Path where cell IDs identified as doublets
will be written.}
}
\value{
invisible(final_results list. Contains multiple items:
\itemize{
\item qc: seurat object prior to filtering and analysis.
\item seu: processed seurat object.
\item sample_qc_pct: Gene expression matrix of union of top N expressed genes in
all samples.
\item clusters_pct: A data frame with expression levels for query genes in each
cluster (or cell type, if annotated).
\item query_exp: A data frame with expression levels for query genes in each
sample.
\item query_pct: A data frame with expression levels for query genes in each
sample.
\item query_cluster_pct: Same as query_pct, but subset by one or even two
conditions.
\item markers: Data frame of marker genes.
\item SingleR_annotation: Trained SingleR annotation object.
\item DEG_list: List of differentially expressed genes across speficied
conditions. Performed globally and in each cluster or cell type
subset.
\item subset_seu: Analyzed seurat objects, subset by integration condition.
\item subset_DEGs: Same as DEG_list, but subset by integration condition.
\item integrate: Whether or not to trigger integrative analysis. Workflow
contains a few key differences depending on this argument.
}
}
\description{
main_annotate_sc
is the main seurat analysis function, which can be used for integrative or
non-integrative analysis.
}
\details{
subset_seu and subset_DEGs only trigger if two conditions have
been provided for integrative analysis, and they will contain two objects,
each only containing samples corresponding to one of the two possible values
in the second experimental condition. That allows isolating the effect
of each experimental condition to be analyzed and compared separately.)
}
\examples{
 \dontrun{
   main_annotate_sc(seu = seurat_object, minqcfeats = 500, percenmt = 5,
                  query = "TREM2", sigfig = 2, resolution = 0.5,
                  p_adj_cutoff = 5e-3, name = "project_name",
                  integrate = TRUE, cluster_annotation = NULL,
                  cell_annotation = cell_types, DEG_target = "genotype",
                  scalefactor = 10000, hvgs = 2000, subset_by = "genotype",
                  normalmethod = "LogNormalize", ndims = 10, verbose = FALSE,
                  output = getwd(), save_RDS = FALSE, reduce = FALSE,
                  ref_label = NULL, SingleR_ref = NULL, ref_de_method = NULL,
                  ref_n = NULL, BPPARAM = NULL, doublet_list = NULL,
                  integration_method = "Harmony", sketch_pct = 12,
                  DEG_p_val_cutoff = 5e-3,
                  min_avg_log2FC = 0.5, k_weight = 90)
 }
}
