% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_degenes_Hunter.R
\name{main_degenes_Hunter}
\alias{main_degenes_Hunter}
\title{Main DEgenes Hunter Function}
\usage{
main_degenes_Hunter(
  raw = NULL,
  target = NULL,
  external_DEA_data = NULL,
  output_files = getwd(),
  reads = 2,
  minlibraries = 2,
  filter_type = "separate",
  p_val_cutoff = 0.05,
  lfc = 1,
  modules = "DELNW",
  minpack_common = 4,
  model_variables = "",
  numerics_as_factors = TRUE,
  custom_model = FALSE,
  string_factors = "",
  numeric_factors = "",
  WGCNA_memory = 5000,
  WGCNA_norm_method = "DESeq2",
  WGCNA_deepsplit = 2,
  WGCNA_min_genes_cluster = 20,
  WGCNA_detectcutHeight = 0.995,
  WGCNA_mergecutHeight = 0.25,
  WGCNA_all = FALSE,
  WGCNA_blockwiseNetworkType = "signed",
  WGCNA_blockwiseTOMType = "signed",
  WGCNA_minCoreKME = 0.5,
  WGCNA_minCoreKMESize = NULL,
  WGCNA_minKMEtoStay = 0.2
)
}
\arguments{
\item{raw}{Dataframe with raw counts per sample and feature}

\item{target}{set with control and treatment information}

\item{external_DEA_data}{external DEA set}

\item{output_files}{otput files path}

\item{reads}{set of reads}

\item{minlibraries}{minimum of libraries to use a set of data}

\item{filter_type}{filter type to ba applied. Allowed: "separate" or "global"}

\item{p_val_cutoff}{p-value threshold}

\item{lfc}{minimum logFold Change}

\item{modules}{modules to be executed. Allowed: DESeq2 (D), NOISeq (N), Limma (L), EdgeR (E), WGCNA (W)}

\item{minpack_common}{minimum of pack that must be significant to tag a gene as significant}

\item{model_variables}{custom model}

\item{numerics_as_factors}{transform numeric values to factors. Default: TRUE}

\item{custom_model}{boolean of usage of custom model. Default: FALSE}

\item{string_factors}{string factors for WGCNA}

\item{numeric_factors}{numeric factors for WGCNA}

\item{WGCNA_memory}{see WGCNA package}

\item{WGCNA_norm_method}{see WGCNA package}

\item{WGCNA_deepsplit}{see WGCNA package}

\item{WGCNA_min_genes_cluster}{see WGCNA package}

\item{WGCNA_detectcutHeight}{see WGCNA package}

\item{WGCNA_mergecutHeight}{see WGCNA package}

\item{WGCNA_all}{see WGCNA package}

\item{WGCNA_blockwiseNetworkType}{see WGCNA package}

\item{WGCNA_blockwiseTOMType}{see WGCNA package}

\item{WGCNA_minCoreKME}{see WGCNA package}

\item{WGCNA_minCoreKMESize}{see WGCNA package}

\item{WGCNA_minKMEtoStay}{see WGCNA package}
}
\value{
expression analysis result object with studies performed
}
\description{
This function allows you to perform the DEG analysis with different algorithms.
}
\examples{
data(toc)
data(target)
degh_out <- main_degenes_Hunter(raw=toc, target=target, modules="D")
}
\keyword{method}
