% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functional_hunter_new.R
\name{main_functional_hunter}
\alias{main_functional_hunter}
\title{Perform Functional analysis with different
enrichment corpora.}
\usage{
main_functional_hunter(
  hunter_results,
  model_organism,
  annot_table = NULL,
  organisms_table = get_organism_table(),
  input_gene_id = "ENSEMBL",
  custom = NULL,
  enrich_dbs = c("MF", "Reactome"),
  enrich_methods = "ORA",
  annotation_source = "orgdb",
  save_query = FALSE,
  pthreshold = 0.1,
  qthreshold = 0.2,
  cores = 1,
  task_size = 1,
  output_files = "results",
  fc_colname = "mean_logFCs"
)
}
\arguments{
\item{hunter_results}{list with DEG analysis results}

\item{model_organism}{organisms which genes are being studied}

\item{annot_table}{(OPTIONAL) annotation table to translate gene IDs}

\item{organisms_table}{(OPTIONAL) configuration table for given organism.
Use see get_organism_table()}

\item{input_gene_id}{(OPTIONAL) type og gene IDs given. Allowed: ENSEMBL (E)
, entrezgene (e), TAIR/Arabidopsis (T), Gene Names (G).}

\item{custom}{(OPTIONAL) list of dataframes with GMT corpus}

\item{save_query}{(OPTIONAL) flag to indicate if query must be stored}

\item{pthreshold}{pvalue threshold}

\item{qthreshold}{qvalue threshold}

\item{cores}{cores to be used if parallel features are going to be used.
Default: 1}

\item{task_size}{number of elements per packages used}

\item{output_files}{output folder}

\item{fc_colname}{main logFC colname (into hunter_results dataframe)}

\item{func_annot_db}{(OPTIONAL) modules to be enriched}

\item{GO_subont}{(OPTIONAL) specific GO subontologies to be enriched}

\item{analysis_type}{(OPTIONAL) enrichment type to be performed. Allowed:
ORA (o) and GSEA (g)}

\item{remote}{(OPTIONAL) remote allowed actions. Are: Biomart query (b) and
KEGG enrichment (k)}
}
\value{
functional result object with enrichments performed
}
\description{
Perform Functional analysis with different
enrichment corpora.
}
\examples{
# Load DE analysis result
data(degh_output)
annots <- "" # Include here wanted ontologies to be used
func_results <- functional_hunter(hunter_results = degh_output,
model_organism = "Mouse",func_annot_db = annots)
}
\keyword{method}
