% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{main_sc_Hunter}
\alias{main_sc_Hunter}
\title{main_sc_Hunter
\code{main_sc_Hunter} is the main seurat analysis function. Can be used
for integrative or non-integrative analysis.}
\usage{
main_sc_Hunter(
  DEG_target,
  seu,
  p_val_cutoff = 0.001,
  min_avg_log2FC = 0.5,
  min_counts = 0.5,
  query = NULL,
  verbose = FALSE,
  min_cell_proportion = 0.01,
  output_path = getwd(),
  top = NULL
)
}
\arguments{
\item{DEG_target}{A string vector. Categories by which DEG analysis will be
performed}

\item{seu}{Seurat object to analyze.}

\item{p_val_cutoff}{Adjusted p-val cutoff for significant DEGs.
Default 5e-3.}

\item{min_avg_log2FC}{Average log2fc cutoff for significant DEGs.
Default 0.5.}

\item{min_counts}{Min counts to consider a gene is expressed in a cell.}

\item{query}{A string vector. List of genes to focus on DEG analysis in
addition to regular DEG analysis.}

\item{verbose}{Print extra execution information.}

\item{min_cell_proportion}{Minimum threshold of percentage of cells
expressing DEG. Despite Seurat::FindMarkers being called "min.pct", it is
not a percentage but a value from 0 to 1.}

\item{output_path}{Path where output will be written.}

\item{top}{An integer. Top N DEGs to represent in certain plots.}
}
\value{
A list containing DEG results, metrics and a query subset of
DEG results.
}
\description{
main_sc_Hunter
\code{main_sc_Hunter} is the main seurat analysis function. Can be used
for integrative or non-integrative analysis.
}
\examples{
data(pbmc_tiny)
seu <- pbmc_tiny
seu$sample <- "a"
seu$sample[8:15] <- "b"
seu$seurat_clusters <- c("0", "1")
DEG_target <- data.frame(sample = seu$sample,
                         treat = c(rep("Ctrl", 7), rep("Treat", 8)))
DEG_results <- main_sc_Hunter(DEG_target = DEG_target, seu = seu,
                        p_val_cutoff = 1, min_avg_log2FC = 0, min_counts = 0,
                        min_cell_proportion = 0)
print(DEG_results)
}
