% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{main_sc_Hunter}
\alias{main_sc_Hunter}
\title{main_sc_Hunter
\code{main_sc_Hunter} is the main seurat analysis function. Can be used
for integrative or non-integrative analysis.}
\usage{
main_sc_Hunter(
  seu = stop("No seurat object supplied."),
  DEG_target = stop("No DEG target supplied"),
  p_val_cutoff = 0.001,
  min_avg_log2FC = 0.5,
  min_counts = 10,
  verbose = FALSE,
  query = NULL,
  min_cell_proportion = 1,
  subset_by = NULL,
  output_path = getwd()
)
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{DEG_target}{A string vector. Categories by which DEG analysis will be
performed}

\item{p_val_cutoff}{Adjusted p-val cutoff for significant DEGs.
Default 5e-3.}

\item{min_avg_log2FC}{Average log2fc cutoff for significant DEGs.
Default 0.5.}

\item{min_counts}{Min counts to consider a gene is expressed in a cell.}

\item{verbose}{Print extra execution information.}

\item{min_cell_proportion}{Minimum threshold of percentage of cells
expressing DEG. Despite Seurat::FindMarkers being called "min.pct", it is
not a percentage but a value from 0 to 1.}

\item{output_path}{Path where output will be written.}
}
\description{
main_sc_Hunter
\code{main_sc_Hunter} is the main seurat analysis function. Can be used
for integrative or non-integrative analysis.
}
