% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{merge_bam_stats}
\alias{merge_bam_stats}
\title{Function to merge bam stats.
\code{merge_bam_stats} Merges the bam stat files contained in the specified path.}
\usage{
merge_bam_stats(stats_path)
}
\arguments{
\item{stats_path}{Bam stats directories of each sample. Stats generated
with samtools idxstats (see Aberrant_Expression workflow for how to generate
this file).}
}
\value{
A table containing the merged bam stats.
}
\description{
Function to merge bam stats.
\code{merge_bam_stats} Merges the bam stat files contained in the specified path.
}
