% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library_new.R
\name{multienricher_gsea}
\alias{multienricher_gsea}
\title{Perform gsea enrichment analysis of a list of genes}
\usage{
multienricher_gsea(
  all_funsys = NULL,
  genes_list,
  organism_info,
  org_db = NULL,
  task_size = 1,
  workers = 1,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  kegg_file = NULL,
  custom_sets = NULL,
  readable = FALSE,
  ...
)
}
\description{
Perform gsea enrichment analysis of a list of genes
}
