% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library_new.R
\name{multienricher_ora}
\alias{multienricher_ora}
\title{Perform ORA enrichment analysis of a list of genes}
\usage{
multienricher_ora(
  all_funsys = NULL,
  genes_list,
  universe = NULL,
  organism_info,
  org_db = NULL,
  task_size = 1,
  workers = 1,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  pAdjustMethod = "BH",
  kegg_file = NULL,
  custom_sets = NULL,
  readable = FALSE,
  ...
)
}
\description{
Perform ORA enrichment analysis of a list of genes
}
