% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_parse_pseudocounts.R
\name{parse_pseudocounts}
\alias{parse_pseudocounts}
\title{main_parse_pseudocounts
Functions that loads pseudommappers results and complite it into a table of counts}
\usage{
parse_pseudocounts(input_folder, mapper, output_type, annotation_file)
}
\arguments{
\item{input_folder}{folder with pseudocounts files}

\item{mapper}{string pseudomapping program}

\item{output_type}{string indicate i fthe output must be 'G' genes, 'T' transcripts or both}

\item{annotation_file}{gtf or gff path, requested for parsin salmon results}
}
\description{
main_parse_pseudocounts
Functions that loads pseudommappers results and complite it into a table of counts
}
\examples{
 \dontrun{
     parse_pseudocounts(input_folder = "Path_to_folder",
                         mapper = "salmon",
                         annotation_file = "path_to_gtf")    
 }
}
\keyword{method}
