% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{preprocess_gtf}
\alias{preprocess_gtf}
\title{Preprocess GTF for DROP analysis}
\usage{
preprocess_gtf(gtf, outdir)
}
\arguments{
\item{gtf}{GTF file to process.}

\item{outdir}{Processed files directory.}
}
\description{
\code{preprocess_gtf} takes a GTF file and a directory path. It outputs a txdb,
count ranges, and gene_mapping files, which it writes to provided directory.
}
\examples{
gtf <- system.file("extData/testdata", "gencode.v45.toy.annotation.gtf",
                    package = "ExpHunterSuite")
}
