% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{process_doublets}
\alias{process_doublets}
\title{process_sketch}
\usage{
process_doublets(seu, name = NULL, qc, doublet_path = getwd())
}
\arguments{
\item{seu}{Seurat object to analyze.}

\item{name}{Sample name.}

\item{qc}{QC object to tag}

\item{doublet_path}{Path where cell IDs identified as doublets will be
written.}
}
\value{
A list containing the updated seurat object and tagged qc object.
}
\description{
\code{process_sketch} simplifies doublet detection and tagging process, as well
as removes doublets from seurat object.
}
\examples{
 \dontrun{
   process_doublets(seu = sample_seu, name = "sample1", qc = qc)
 }
}
