% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{process_sc_params}
\alias{process_sc_params}
\title{process_sc_params}
\usage{
process_sc_params(params = list(), mode = "annotation")
}
\arguments{
\item{params}{A list of parameters to parse.}

\item{mode}{A string. Scripting mode to use. Has special behaviours for
"annotation" (the default) and "DEG". Set to NULL or any other value for
query mode. For coherence, we set it to "query" in our query script.}
}
\value{
A list containing parsed options.
}
\description{
\code{process_sc_params} processes parameter list and turns it into an option list
usable by main sc functions.
}
\examples{
 \dontrun{
   process_sc_params(params = annot_params, mode = "annotation")
 }
}
