% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{process_sketch}
\alias{process_sketch}
\title{process_sketch}
\usage{
process_sketch(
  seu,
  sketch_method,
  sketch_pct,
  force_ncells,
  hvgs,
  verbose = FALSE,
  features = Seurat::VariableFeatures(seu)
)
}
\arguments{
\item{seu}{Seurat object to sketch.
Default value of 5000, recommended by Seurat tutorials.}

\item{sketch_method}{A string. Method to use in score calculation for
sketching.}

\item{sketch_pct}{A numeric. Percentage of total cells to consider
representative of the experiment. Default 12, as suggested by sketching
tutorial.}

\item{force_ncells}{an integer. Sketch function will be forced to select
this many cells.}

\item{hvgs}{An integer. Number of highly-variable features to select.
default 2000.}

\item{verbose}{A boolean, also passed to Seurat::FindVariableFeatures call.
\itemize{
\item \code{TRUE}: Will print out execution time, in addition to passing its value
to Seurat::FindVariableFeatures call..
\item \code{FALSE} (the default): Function will be quiet.
}}

\item{features}{A character vector of feature names to include in the
sketched assay.}
}
\value{
A seurat object with a new sketched assay.
}
\description{
\code{process_sketch} is a wrapper for Seurat sketching step.
}
\examples{
 \dontrun{
   process_sketch(seu = seu, sketch_method = sketch_method,
                  sketch_pct = sketch_pct, force_ncells = force_ncells,
                  hvgs = hvgs, verbose = verbose)
 }
}
