% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{project_sketch}
\alias{project_sketch}
\title{project_sketch
is a wrapper for the sketch projection steps.}
\usage{
project_sketch(seu, reduction, ndims)
}
\arguments{
\item{seu}{Seurat object whose sketch data to project.}

\item{reduction}{Dimensional reduction name for batch-corrected embeddings
in the sketched object (default is 'integrated_dr')}

\item{ndims}{Dimensions to select in ProjectData.}
}
\value{
Seurat object containing sketch projection in RNA assay.'
}
\description{
project_sketch
is a wrapper for the sketch projection steps.
}
\examples{
 \dontrun{
   project_sketch(seu = seu, reduction = "umap", ndims = 10)
 }
}
