% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{read_and_format_targets}
\alias{read_and_format_targets}
\title{read_and_format_targets
\code{read_and_format_targets} formats a marker-celltype table into a list}
\usage{
read_and_format_targets(file)
}
\arguments{
\item{file}{Path to target gene file}
}
\value{
A list with one element per cell type
}
\description{
read_and_format_targets
\code{read_and_format_targets} formats a marker-celltype table into a list
}
