% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{read_sc_counts}
\alias{read_sc_counts}
\title{read_sc_counts}
\usage{
read_sc_counts(
  name,
  input,
  mincells = 1,
  minfeats = 1,
  exp_design,
  lower = 100,
  niters = 10000,
  test.ambient = FALSE,
  ignore = NULL,
  alpha = NULL,
  round = TRUE,
  by.rank = NULL,
  known.empty = NULL,
  BPPARAM = SerialParam
)
}
\arguments{
\item{name}{sample name}

\item{input}{path to cellranger counts}

\item{mincells}{min number of cells for which a feature is recorded}

\item{minfeats}{min number of features for which a cell is recorded}

\item{exp_design}{experiment design table}
}
\value{
Seurat object
}
\description{
\code{read_sc_counts} creates a seurat object from cellranger counts
}
\examples{
 \dontrun{
   read_sc_counts(name = "experiment_name", mincells = 1, minfeats = 1,
                  input = "path\to\counts\matrices",
                  exp_design = exp_design)
 }
}
