% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DROP_functions.R
\name{run_outrider}
\alias{run_outrider}
\title{Wrapper for DROP main OUTRIDER script.
\code{run_outrider} takes a filtered Outrider dataset and runs the OUTRIDER
algorithm on it.}
\usage{
run_outrider(ods_unfitted, implementation, max_dim_proportion)
}
\arguments{
\item{ods_unfitted}{The filtered Outrider dataset on which to run OUTRIDER.}

\item{implementation}{Method for sample covariation removal in OUTRIDER.
Possible values: "autoencoder" (default), "pca" or "peer".}

\item{max_dim_proportion}{Maximum value for autoencoder encoding
dimension. Default 3, optimum for aberrant expression.}
}
\value{
A fitted outrider dataset.
}
\description{
Wrapper for DROP main OUTRIDER script.
\code{run_outrider} takes a filtered Outrider dataset and runs the OUTRIDER
algorithm on it.
}
