% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{run_scDblFinder}
\alias{run_scDblFinder}
\title{\code{run_scDblFinder} is a wrapper for the scDblFinder pipeline}
\usage{
run_scDblFinder(seu, assay, includePCs, nfeatures)
}
\arguments{
\item{seu}{Input seyrat object.}

\item{includePCs}{The index of principal components to include in the
predictors (e.g. `includePCs=1:2`), or the number of top components to use
(e.g. `includePCs=10`, equivalent to 1:10).}

\item{nfeatures}{The number of top features to use. Alternatively, a 
character vectors of feature names (e.g. highly-variable genes) to use.}
}
\value{
A list. Item "seu" contains seurat object with tagged doublets
in metadata slot. Item "barcodes" contains a vector of cell barcodes
corresponding to barcodes.
}
\description{
\code{run_scDblFinder} is a wrapper for the scDblFinder pipeline
}
\examples{
\dontrun{
   run_scDblFinder(seu, "counts", 10, 2000)
}
}
