% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis_library.R
\name{scale_data_matrix}
\alias{scale_data_matrix}
\title{Scale a matrix using minimum-maximum method}
\usage{
scale_data_matrix(data_matrix, transpose = FALSE)
}
\arguments{
\item{data_matrix}{to be scaled}

\item{transpose}{boolen flag which indicates if matrix should be transposed.
Default: FALSE}
}
\value{
scaled matrix
}
\description{
Scale a matrix using minimum-maximum method
}
\keyword{method}
