% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{sketch_sc_experiment}
\alias{sketch_sc_experiment}
\title{sketch_sc_experiment
determines optimal cell number to sketch seurat assay. If it is larger than
specified minimum, it proceeds with sketching, else it is skipped.}
\usage{
sketch_sc_experiment(
  seu,
  min.ncells = 5000,
  assay = "RNA",
  method = "LeverageScore",
  sketched.assay = "sketch",
  cell.pct = 12,
  force.ncells = NA_integer_
)
}
\arguments{
\item{seu}{Seurat object to sketch.}

\item{min.ncells}{An integer. If estimated cell number optimal value for
sketching is smaller than this number, sketching will not be performed.
Default value of 5000, recommended by Seurat tutorials.}

\item{assay}{A string. Name of assay to sketch. Default "RNA".}

\item{method}{A string. Method to use for sketching. Default "LeverageScore".}

\item{sketched.assay}{A string. Name of assay where sketch will be saved.}

\item{cell.pct}{A numeric. Percentage of total cells to consider
representative of the experiment. Default 12, as suggested by sketching
tutorial.}

\item{force.ncells}{An integer. Number of cells to forcibly use in sketch.
NA by default.}
}
\value{
A seurat object with a new sketched assay.
}
\description{
sketch_sc_experiment
determines optimal cell number to sketch seurat assay. If it is larger than
specified minimum, it proceeds with sketching, else it is skipped.
}
\examples{
 \dontrun{
   sketched_experiment <- sketch_sc_experiment(seu = unsketched_experiment,
          assay = "RNA", method = "LeverageScore", sketched.assay = "sketch",
          cell.pct = 50)
 }
}
