% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{subset_seurat}
\alias{subset_seurat}
\title{subset_seurat
\code{subset_seurat} subsets a seurat object by a specified value of provided
column.}
\usage{
subset_seurat(seu, column, value)
}
\arguments{
\item{seu}{Seurat object}

\item{column}{Column with value by which to subset input}

\item{value}{Value to search within column}
}
\value{
A subset of the seurat object, which itself is a seurat object.
}
\description{
subset_seurat
\code{subset_seurat} subsets a seurat object by a specified value of provided
column.
}
\examples{
data(pbmc_tiny)
head(subset_seurat(seu = pbmc_tiny, column = "groups", value = "g2"))
}
