% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{tag_DEGs}
\alias{tag_DEGs}
\title{Tag DEGs by failed filter}
\usage{
tag_DEGs(
  DEG_df,
  p_val_cutoff = 0.1,
  min_avg_log2FC = 0.5,
  min_cell_proportion = 0.1
)
}
\arguments{
\item{DEG_df}{DEG data frame to tag.}

\item{p_val_cutoff}{Genes with an adjusted P-value higher than this argument
will be tagged as high_pval.}

\item{min_avg_log2FC}{Genes whose average log2FC is smaller than this value
will be tagged as low_fc.}

\item{min_cell_proportion}{Genes expressed in a percentage of cells smaller
than this number in any of the two groups will be tagged as low_cell_pct.}
}
\description{
\code{tag_DEGs} takes a DEG data frame and adds tags according to several
filters.
}
\examples{
DEGs <- data.frame(avg_log2FC = c(0, 0.2, 1, -0.1),
                   gene = c("PPBP", "IGLL5", "VDAC3", "GNLY"),
                   p_val_adj = c(0, 0, 0, 0), pct.1 = c(0, 0, 0.5, 0.5),
                   p_val_adj = c(1, 0, 0.5, 0.1), pct.2 = c(0, 0.5, 0, 0.5),
                   avg_log2FC = c(10, 5, 0, 0.5))
DEGs <- tag_DEGs(DEG_df = DEGs)
}
