% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{tag_qc}
\alias{tag_qc}
\title{tag_gc
Perform Quality Control}
\usage{
tag_qc(seu, minqcfeats, percentmt)
}
\arguments{
\item{seu}{Seurat object}

\item{minqcfeats}{Min number of features for which a cell is selected}

\item{percentmt}{Max percentage of reads mapped to mitochondrial genes for which a cell is selected}
}
\value{
Seurat object
}
\description{
tag_gc
Perform Quality Control
}
\keyword{preprocessing,}
\keyword{qc}
