% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_library.R
\name{tag_qc}
\alias{tag_qc}
\title{tag_qc}
\usage{
tag_qc(
  seu,
  minqcfeats = 500,
  percentmt = 5,
  doublet_list = NULL,
  min_cells_per_sample = 500
)
}
\arguments{
\item{seu}{Seurat object to tag}

\item{minqcfeats}{An integer. Minimum features to consider a cell valid.
Default 500}

\item{percentmt}{A float. Maximum MT percentage to consider a cell valid.
Default 5}

\item{doublet_list}{Vector of barcodes to be marked as doublet. Default NULL.}

\item{min_cells_per_sample}{An integer. Samples with fewer than this value of
cells will be tagged as low_cell.'}
}
\value{
Seurat object with tagged metadata
}
\description{
\code{tag_qc} tags cell IDs not passing QC filters
}
\examples{
seu <- pbmc_tiny
seu$sample <- c("a", "b", "c")
seu$sample[1] <- "c"
# Integrative experiment, previously loaded doublet list:
head(tag_qc(seu = seu, minqcfeats = 500, percentmt = 5,
            doublet_list = c("ATGCCAGAACGACT", "CATGGCCTGTGCAT")))
# Per-sample experiment, no prior doublet knowledge
head(tag_qc(seu = seu, minqcfeats = 1, percentmt = 5, doublet_list = NULL,
            min_cells_per_sample = 5)) 
}
\keyword{preprocessing,}
\keyword{qc}
