% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{write_DEG_output}
\alias{write_DEG_output}
\title{Write DEG results to disk
\code{write_DEG_output }writes DEG matrices of a single-cell experiment, along
with metrics and metadata.'}
\usage{
write_DEG_output(name, DEG_list, opt = NULL)
}
\arguments{
\item{name}{Name of DEG_list element to process.}

\item{DEG_list}{Complete list of DEG results.}

\item{opt}{Options used in script call.}
}
\value{
invisible(NULL)
}
\description{
Write DEG results to disk
\code{write_DEG_output }writes DEG matrices of a single-cell experiment, along
with metrics and metadata.'
}
\examples{
\dontrun{
   write_DEG_output(DEG_results = test_results, opt = params,
                    name = "important_comparison")
}
}
