% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{write_dense_matrix}
\alias{write_dense_matrix}
\title{write_dense_matrix
write dense matrix to a format compatible with our workflow.}
\usage{
write_dense_matrix(counts, output)
}
\arguments{
\item{counts}{A dense matrix.}

\item{output}{Name of rds file to save.}
}
\value{
invisible(NULL)
}
\description{
write_dense_matrix
write dense matrix to a format compatible with our workflow.
}
\details{
Under some specific cases, scaled counts data are created in dense format,
and we have been unable to find a way to write them to MatrixMarket format.
For now, this function merely saves them in rds format. Hopefully we will
eventually find a way to properly handle this case and save it in the right
format.
}
\examples{
\dontrun{
   write_dense_matrix(counts = matrix, output = "./test.rds")
}
}
