% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_functional_report}
\alias{write_functional_report}
\title{Write Main DEgenes Hunter functional report
This function allows you to report the Functional analysis.}
\usage{
write_functional_report(
  hunter_results,
  func_results,
  cores = 2,
  output_files = getwd(),
  organisms_table = NULL,
  fc_colname = "mean_logFCs",
  task_size = 1,
  report = "fc",
  source_folder = NULL,
  group_results = FALSE,
  showCategories = 30,
  corr_threshold = 0.8,
  pvalcutoff = 0.05,
  template_folder = NULL,
  max_genes = 200,
  node_label = "category",
  size_item = 1,
  size_category = 1,
  size_edge = 1,
  hilight = "none",
  hilight_alpha = 0.3,
  files_css = file.path(template_folder, "styles.css")
)
}
\arguments{
\item{hunter_results}{DEG analysis results}

\item{func_results}{functional results}

\item{cores}{(OPTIONAL) cores for parallel features}

\item{output_files}{output folder.}

\item{organisms_table}{(OPTIONAL) configuration table for given organism.
Use see get_organism_table()}

\item{fc_colname}{(OPTIONAL) main logFC colname (into hunter_results
dataframe)}

\item{task_size}{number of elements per packages used}

\item{report}{string with reports to be written. Allowed: clusters (c)}

\item{source_folder}{htmlreportR source folder to load js and css libraries}

\item{group_results}{experimental - whether to group results in the emap plot
and functional (f). Default = "fc"}

\item{showCategories}{number of categories in the figures (per cluster)}

\item{corr_threshold}{minimum module eigengene-trait vector absolute Pearson
R value}

\item{pvalcutoff}{maximum module eigengene-trait vector correlation P value}

\item{template_folder}{(OPTIONAL) RMD templates folder}

\item{max_genes}{maximum number of genes to plot in cnet plot}

\item{node_label}{How to label the nodes. Default "category". Other possible
values: "all", "none", "item", "exclusive", "share".}

\item{size_item}{Size of nodes. Smaller than 1 values will reduce size,
higher than 1 values will increase it. Default: 1.}

\item{size_category}{Size of category labels. Smaller than 1 values will
reduce size, higher than 1 values will increase it. Default: 1.}

\item{size_edge}{Size of edges. Smaller than 1 values will reduce size,+
higher than 1 values will increase it. Default: 1.}

\item{hilight}{Name of category to highlight. Default "none".}

\item{hilight_alpha}{Transparency value that will be applied to categories
not highlighted. Must be a value between 0 (completely transparent) and 1
(completely opaque). Default: 0.3.}

\item{files_css}{Path to css file to inject. By default it injects styles.css
found in ExpHunterSuite's templates directory.}
}
\value{
void
}
\description{
Write Main DEgenes Hunter functional report
This function allows you to report the Functional analysis.
}
\examples{
\dontrun{
# Load func and DE results
data(degh_output)
func_results <- list() 
func_results <- main_functional_hunter(degh_output, "Mouse")
write_functional_report(degh_output, func_results)
}
}
