% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_functional_report}
\alias{write_functional_report}
\title{Write Main DEgenes Hunter functional report
This function allows you to report the Functional analysis.}
\usage{
write_functional_report(
  hunter_results,
  func_results,
  cores = 2,
  output_files = getwd(),
  organisms_table = NULL,
  fc_colname = "mean_logFCs",
  task_size = 1,
  report = "fc",
  source_folder = NULL,
  group_results = FALSE,
  showCategories = 30,
  corr_threshold = 0.8,
  pvalcutoff = 0.05,
  template_folder = NULL,
  max_genes = 200
)
}
\arguments{
\item{hunter_results}{DEG analysis results}

\item{func_results}{functional results}

\item{cores}{(OPTIONAL) cores for parallel features}

\item{output_files}{output folder.}

\item{organisms_table}{(OPTIONAL) configuration table for given organism.
Use see get_organism_table()}

\item{fc_colname}{(OPTIONAL) main logFC colname (into hunter_results
dataframe)}

\item{task_size}{number of elements per packages used}

\item{report}{string with reports to be written. Allowed: clusters (c)}

\item{source_folder}{htmlreportR source folder to load js and css libraries}

\item{group_results}{experimental - whether to group results in the emap plot
and functional (f). Default = "fc"}

\item{showCategories}{number of categories in the figures (per cluster)}

\item{corr_threshold}{minimum module eigengene-trait vector absolute Pearson
R value}

\item{pvalcutoff}{maximum module eigengene-trait vector correlation P value}

\item{template_folder}{(OPTIONAL) RMD templates folder}

\item{max_genes}{maximum number of genes to plot in cnet plot}
}
\value{
void
}
\description{
Write Main DEgenes Hunter functional report
This function allows you to report the Functional analysis.
}
\examples{
/dontrun{
# Load func and DE results
data(degh_output)
func_results <- list() 
func_results <- main_functional_hunter(degh_output, "Mouse")
write_functional_report(degh_output, func_results)
}
}
