% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_global_cormit}
\alias{write_global_cormit}
\title{Write global coRmiT report}
\usage{
write_global_cormit(
  cormit_res,
  output_files = getwd(),
  template_folder = NULL,
  source_folder = NULL,
  files_css = file.path(template_folder, "styles.css")
)
}
\arguments{
\item{cormit_res}{coRmiT results object}

\item{output_files}{output file path}

\item{template_folder}{htmlreportR templates folder}

\item{source_folder}{Folder where js and css templates to load in htmlreportR
object are located.}

\item{files_css}{Path to css file to inject. By default it injects styles.css
found in ExpHunterSuite's templates directory.}
}
\value{
void
}
\description{
This function writes a report of coRmiT results.
}
\examples{
# Load DE analysis results
\dontrun{
     degh_output <- list() # data(degh_output)
     write_global_cormit(degh_output)
}
}
