% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_merged_cluster_report}
\alias{write_merged_cluster_report}
\title{Write clustering report for functional Hunter
This function will usually be called by write_functional_report.}
\usage{
write_merged_cluster_report(
  enrichments_ORA,
  results_path,
  template_folder,
  sample_classes = NULL,
  DEGH_results = NULL,
  n_category,
  node_label = "category",
  size_item = 1,
  size_category = 1,
  size_edge = 1,
  hilight = "none",
  hilight_alpha = 0.3,
  group_results,
  func_results = NULL,
  source_folder = NULL,
  top_categories,
  files_css = file.path(template_folder, "styles.css")
)
}
\arguments{
\item{enrichments_ORA}{ORA enrichment resunts.}

\item{results_path}{Path where report will be written.}

\item{template_folder}{(OPTIONAL) RMD templates folder}

\item{sample_classes}{Sample classes.}

\item{DEGH_results}{DEGenes Hunter results object.}

\item{n_category}{An integer. Number of categories to pass as showCategory
value to several clusterProfiler plotting functions.}

\item{node_label}{How to label the nodes. Default "category". Other possible
values: "all", "none", "item", "exclusive", "share".}

\item{size_item}{Size of nodes. Smaller than 1 values will reduce size,
higher than 1 values will increase it. Default: 1.}

\item{size_category}{Size of category labels. Smaller than 1 values will
reduce size, higher than 1 values will increase it. Default: 1.}

\item{size_edge}{Size of edges. Smaller than 1 values will reduce size,+
higher than 1 values will increase it. Default: 1.}

\item{hilight}{Name of category to highlight. Default "none".}

\item{hilight_alpha}{Transparency value that will be applied to categories
not highlighted. Must be a value between 0 (completely transparent) and 1
(completely opaque). Default: 0.3.}

\item{group_results}{experimental - whether to group results in the emap plot
and functional (f). Default = "fc"}

\item{func_results}{functional results}

\item{source_folder}{htmlreportR source folder to load js and css libraries}

\item{files_css}{Path to css file to inject. By default it injects styles.css
found in ExpHunterSuite's templates directory.}
}
\value{
invisible(NULL)
}
\description{
Write clustering report for functional Hunter
This function will usually be called by write_functional_report.
}
