% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_merged_cluster_report}
\alias{write_merged_cluster_report}
\title{Write clustering report for functional Hunter
This function will usually be called by write_functional_report.}
\usage{
write_merged_cluster_report(
  enrichments_ORA,
  results_path,
  template_folder,
  sample_classes = NULL,
  DEGH_results = NULL,
  showCategories,
  group_results,
  func_results = NULL,
  source_folder = NULL
)
}
\arguments{
\item{template_folder}{(OPTIONAL) RMD templates folder}

\item{showCategories}{number of categories in the figures (per cluster)}

\item{group_results}{experimental - whether to group results in the emap plot
and functional (f). Default = "fc"}

\item{func_results}{functional results}

\item{source_folder}{htmlreportR source folder to load js and css libraries}
}
\description{
Write clustering report for functional Hunter
This function will usually be called by write_functional_report.
}
