% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sc_functions.R
\name{write_sc_report}
\alias{write_sc_report}
\title{write_sc_report
Write integration HTML report}
\usage{
write_sc_report(
  final_results,
  output = getwd(),
  template_folder,
  source_folder = NULL,
  query = NULL,
  subset_by = NULL,
  opt,
  cell_annotation = NULL,
  template = NULL,
  out_name = NULL,
  use_canvas = TRUE,
  extra_columns = NULL,
  DEG_p_val_cutoff = NA_real_,
  min_avg_log2FC = NA_real_
)
}
\arguments{
\item{final_results}{Output from main_sc_Hunter function}

\item{output}{directory where report will be saved}

\item{template_folder}{directory where template is located}

\item{source_folder}{htmlreportR source folder}

\item{query}{A string vector. List of genes to explore in dataset}

\item{subset_by}{factors present in experiment design}

\item{cell_annotation}{A data frame. Table of markers to use for cell type
annotation.}

\item{template}{Template to render}

\item{out_name}{suffix to add to output file name. Useful when rendering
different templates with this function (as is the case in our workflow)}

\item{use_canvas}{Parameter to select whether or not CanvasXpress plots will
be triggered in templates where this control parameter has been implemented.
Setting it to FALSE can be useful for big datasets, as CanvasXpress might
have trouble in certain plots}

\item{name}{experiment name, will be used to build output file name}
}
\value{
invisible(NULL)
}
\description{
write_sc_report
Write integration HTML report
}
\keyword{preprocessing,}
\keyword{report}
\keyword{write,}
